### 2019-09-16 6bf8059
Add Dorefa quantizer
