### 2019-09-16 cc0f8e3
Black Formatting
