### 2019-09-17 92a68cf
Created CompatibilityMetric class, added register_alias decorator to utils
