### 2019-09-17 ec16ac4
Update graph of magnitude aware sign in documentation (#224)
