### 2019-09-17 68b46eb
added boolean is_weight_metric to FlipRatio so we can distinguish weight and activation metrics
