### 2019-09-18 0608a10
update tests
