### 2019-09-18 #1746 @whedon
[REVIEW]: Larq: An Open-Source Library for Training Binarized Neural Networks
**Submitting author:** @lgeiger (<a href="http://orcid.org/0000-0002-8697-9920">Lukas Geiger</a>)
**Repository:** <a href="https://github.com/larq/larq" target ="_blank">https://github.com/larq/larq</a>
**Version:** v0.7.1
