### 2019-09-18 b965884
Added default get_config() and reset_state() functions to _CompatibilityMetric
