### 2019-09-18 2ffdf8e
Cleanup `.bib` file and add arXiv URL to paper (#226)
