+++ paper/paper.bib
@@ -23,6 +23,7 @@
     archiveprefix = {arXiv},
     eprint       = {1906.02107},
     primaryclass = {cs.LG},
+    url          = {https://arxiv.org/abs/1906.02107},
     adsurl       = {https://ui.adsabs.harvard.edu/abs/2019arXiv190602107H},
     adsnote      = {Provided by the SAO/NASA Astrophysics Data System}
 }
@@ -37,6 +38,5 @@
     archiveprefix = {arXiv},
     arxivid      = {1602.02830},
     eprint       = {1602.02830},
-    file         = {:home/koen/Documents/Mendeley-Plumerai/Courbariaux et al. - 2016 - Binarized Neural Networks Training Deep Neural Networks with Weights and Activations Constrained to 1 or -1.pdf:pdf},
     pmid         = 23554596
 }
