### 2019-09-18 81b4d96
Add Dorefa quantizer (#223)
* Add Dorefa quantizer

* Black Formatting

* Feedback#1

* Correct name of quantizer function

* update tests

* c

Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>

* update test

Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
