+++ larq/quantizers.py
@@ -69,6 +69,8 @@ __all__ = [
     "ste_tern",
     "SteHeaviside",
     "ste_heaviside",
+    "DoReFaQuantizer",
+    "dorefa_quantizer",
 ]
 
 
@@ -534,6 +536,93 @@ def ste_heaviside(x, clip_value=1.0):
         return tf.sign(tf.nn.relu(x)), grad
 
     return _and_binarize_with_identity_grad(x)
+
+
+@utils.register_keras_custom_object
+@utils.set_precision(2)
+def dorefa_quantizer(x, k_bit=2):
+    r"""k_bit quantizer as in the DoReFa paper.
+
+    \\[
+    q(x) = \begin{cases}
+    0 & x < \frac{1}{2n} \\\
+    \frac{i}{n} & \frac{2i-1}{2n} < |x| < \frac{2i+1}{2n} \text{ for } i \in \\{1,n-1\\}\\\
+     1 & \frac{2n-1}{2n} < x
+    \end{cases}
+    \\]
+
+    where \\(n = 2^{\text{k_bit}} - 1\\). The number of bits, k_bit, needs to be passed as an argument.
+    The gradient is estimated using the Straight-Through Estimator
+    (essentially the binarization is replaced by a clipped identity on the
+    backward pass).
+    \\[\frac{\partial q(x)}{\partial x} = \begin{cases}
+    1 &  0 \leq x \leq 1 \\\
+    0 & \text{else}
+    \end{cases}\\]
+
+    ```plot-activation
+    quantizers.dorefa_quantizer
+    ```
+
+    # Arguments
+    k_bit: number of bits for the quantization.
+
+    # Returns
+    quantized tensor
+
+    # References
+    - [DoReFa-Net: Training Low Bitwidth Convolutional Neural Networks
+      with Low Bitwidth Gradients](https://arxiv.org/abs/1606.06160)
+    """
+    x = tf.clip_by_value(x, 0.0, 1.0)
+
+    @tf.custom_gradient
+    def _k_bit_with_identity_grad(x):
+        n = 2 ** k_bit - 1
+        return tf.round(x * n) / n, lambda dy: dy
+
+    return _k_bit_with_identity_grad(x)
+
+
+@utils.register_keras_custom_object
+class DoReFaQuantizer(QuantizerFunctionWrapper):
+    r"""Instantiates a serializable k_bit quantizer as in the DoReFa paper.
+
+    \\[
+    q(x) = \begin{cases}
+    0 & x < \frac{1}{2n} \\\
+    \frac{i}{n} & \frac{2i-1}{2n} < |x| < \frac{2i+1}{2n} \text{ for } i \in \\{1,n-1\\}\\\
+     1 & \frac{2n-1}{2n} < x
+    \end{cases}
+    \\]
+
+    where \\(n = 2^{\text{k_bit}} - 1\\). The number of bits, k_bit, needs to be passed as an argument.
+    The gradient is estimated using the Straight-Through Estimator
+    (essentially the binarization is replaced by a clipped identity on the
+    backward pass).
+    \\[\frac{\partial q(x)}{\partial x} = \begin{cases}
+    1 &  0 \leq x \leq 1 \\\
+    0 & \text{else}
+    \end{cases}\\]
+    
+    ```plot-activation
+    quantizers.dorefa_quantizer
+    ```
+
+    # Arguments
+    k_bit: number of bits for the quantization.
+
+    # Returns
+    Quantization function
+
+    # References
+    - [DoReFa-Net: Training Low Bitwidth Convolutional Neural Networks
+      with Low Bitwidth Gradients](https://arxiv.org/abs/1606.06160)
+    """
+
+    def __init__(self, k_bit):
[…]
