### 2019-09-18 #1746 @sbrugman
Hi @lgeiger, thanks for your submission.

I'm doing this review in parts, as for today I don't have time to fully read up on the topic, but I wanted to provide some feedback for you to be able to start. This comment contains those feedback points. For now, I took the documentation and code into consideration, not the paper.

The Larq ecosystem seems to be an excellent contribution to the open-source community and supports a thrilling research direction. In general, I'm positive. The comments below are intended to help improve the clarity, consideration and completeness of the contribution.

### Documentation
> *A statement of need*: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The documentation clearly states that it is designed to solve power consumption and speed issues with deep neural network architectures. A notion of the target audience is mentioned on the 'about' page ("researchers and developers"), 'papers using Larq' page and in the last paragraph of the paper. I suggest clarifying this directly on the homepage.

The documentation and paper make no claims on the comparison of performance between Quantized Neural Networks and their high-precision weight and bias counterparts. At the time of writing, there is clearly a trade-off (although the authors are working hard to close the gap). I suggest being transparent in this difference and possibly opening the option of educating users on when a few percents of accuracy on a specific dataset are essential and when not. This might be especially relevant for people working in the industry.

> *Community*: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

There is a community menu item with a contribution guide tab in the documentation. This clearly states the steps for third parties wishing to contribute to the software (point 1). The other use cases are undocumented. Links to GitHub issues and spectrum are missing in the document.

> *Example usage*: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The authors provide numerous [examples](https://larq.dev/examples/mnist/). There are even [hosted examples](https://mybinder.org/v2/gh/larq/larq/master?filepath=docs%2Fexamples). In general, the examples are well-documented, except for the Larq Zoo. There is minimal context at these [examples](https://larq.dev/models/examples/), which are more like code snippets. I suggest extending the description for these examples somewhat.

### Other points
- For each of the items in the API, one could ask: how does this compare to the underlying framework and why does it need an extension? Most of the sections (e.g. Quantizers) answer these questions, for example by providing references. An example where this information is missing is [sign](https://larq.dev/api/math/). Adding a reference to the [tensorflow sign](https://www.tensorflow.org/api_docs/python/tf/math/sign) and/or mentioning that `tf.math.sign` would be a quick win.

- The FAQ is now located under "Why Larq?", which I found to be an illogical location.

- Inline math is not rendered (ex. [key concepts](https://larq.dev/guides/key-concepts/)).

- Consider adding references to the home page (e.g. Courbariaux).

Feel free to take over any of the suggestions. I will complete the review at a later time.
