### 2019-09-18 e6dca67
no longer explicitly import keras backend since we only use it once
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
