### 2019-09-18 8e9b66e
Update larq/metrics.py
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
