### 2019-09-18 3bfde67
Update larq/metrics.py
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
