+++ docs/about.md
@@ -1,6 +1,6 @@
-Larq is an open-source Binarized Neural Network (BNN) training library on top of TensorFlow, developed and maintained by [Plumerai](https://plumerai.com). We believe efficient inference of deep learning models will radically change the world by enabling a multitude of new technologies - from delivery drones and self-driving cars to fruit picking robots and autonomous Mars rovers.
+Larq is an open-source Binarized Neural Network (BNN) training library on top of TensorFlow, developed and maintained by [Plumerai](https://plumerai.com). We believe efficient inference of deep learning models will radically change the world by enabling a multitude of new technologies - from delivery drones and self-driving cars to fruit-picking robots and autonomous Mars rovers.
 
-BNNs are the key that will unlock this future. These networks have a much smaller memory footprint than ordinary neural networks, and their bitwise nature makes it possible to execute the computations extremely efficienciently.
+BNNs are the key that will unlock this future. These networks have a much smaller memory footprint than ordinary neural networks, and their bitwise nature makes it possible to execute the computations extremely efficiencient.
 
 Larq has been created to make it easier for researchers and developers to explore the possibilities of BNNs. We strongly believe an open approach to knowledge is the way forward, which is why Larq lives under a permissive open-source software [Apache 2.0 licence](https://github.com/larq/larq/blob/master/LICENSE).
 
