### 2019-09-18 0539f30
Update bnn-architecture.md
