### 2019-09-18 a7e97e6
Renamed compatibilityMetric to LarqMetric, removed unnecessary get_config
