### 2019-09-18 #1746 @casperdcl
more feedback

### General checks
- **Contribution and authorship:** @koenhelwegen seems to have made some significant contributions - at least worth an acknowledgement if not co-authorship. Seems to be the first author of the paper (Helwegen et al.) which inspired this work.

### Functionality
- **Installation:** `pip install larq` is not sufficient (https://github.com/larq/larq#installation and https://larq.dev/#installation), see below for more comments
- **Performance:** There is frequent mention of the efficiency of BNNs, however it would be nice to have e.g. a quantitative table of results showing that `larq` matches theoretical/literature speed improvements.
- **Submission requirements:** @terrytangyuan I'm not sure but as per https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements the software seems to be bordering on a "'thin' API client" around `tensorflow.keras`. Perhaps it's worth submitting to JOSE instead? Or including in a larger publication similar to the Helwegen et al. paper? For example `ls larq/*.py larq/snapshots/*.py | grep -v test | xargs cloc` yields 1,586 non-blank/comment lines, with the main `layers.py` file exclusively made up of lines which look like:

```python
def __init__(
    self,
    units,
    activation=None,
    use_bias=True,
    input_quantizer=None,
    kernel_quantizer=None,
    kernel_initializer="glorot_uniform",
    bias_initializer="zeros",
    kernel_regularizer=None,
    bias_regularizer=None,
    activity_regularizer=None,
    kernel_constraint=None,
    bias_constraint=None,
    metrics=None,
    **kwargs,
):
    super().__init__(
        units,
        activation=activation,
        use_bias=use_bias,
        input_quantizer=input_quantizer,
        kernel_quantizer=kernel_quantizer,
        kernel_initializer=kernel_initializer,
        bias_initializer=bias_initializer,
        kernel_regularizer=kernel_regularizer,
        bias_regularizer=bias_regularizer,
        activity_regularizer=activity_regularizer,
        kernel_constraint=kernel_constraint,
        bias_constraint=bias_constraint,
        metrics=metrics,
        **kwargs,
    )
```

### Documentation

- **A statement of need:** https://larq.dev/about/ doesn't mention students (while the paper does)
- **Installation instructions:** Installation (https://github.com/larq/larq#installation and https://larq.dev/#installation) also needs to also document `pip install tensorflow`, `pip install tensorflow-gpu` (or add them to `setup.py` etc), referencing the earlier link (https://www.tensorflow.org/install) for additional dependencies. Hte par
- **Functionality documentation:**: The API link (https://larq.dev/api/layers/) seems to imply this is just `larq.layers`, yet the Models link (https://larq.dev/models/) gives a separate page which seems to be detailing API for `larq/zoo` which seems to be a separate package built on top of the `larq.layers` API. If this is the case it needs to be made clearer.

### Software paper

- **State of the field:** No description of alternatives and/or how this software compares to other commonly-used packages
- **Quality of writing:**
  + There's only a summary section. Consider adding:
    * code snippets
    * a brief overview of the API (https://larq.dev/api/layers/ seems to imply this is just `larq.layers`)
    * links to resources (`larq/zoo`, documentation website, etc.), result graphs/figures comparing to plain non-BNN `keras`.
  + `tf.keras` is a common alias/abbreviation but isn't a module - it's `tensorflow.keras`
  + "well-tested implementations" could do with a reference to CI coverage etc
- **References:**
  + Inconsistent referencing (e.g. Abadi "et al." versus Chollet "& others")
  + In-text author list for Courbariaux, Hubara, Soudry, El-Yaniv, & Bengio seems a little long
  + The DOI for Courbariaux et al. is for a different work by He et al.
