### 2019-09-18 f4b0c79
Update key-concepts.md
