### 2019-09-18 8ccbd3a
Spelling and Grammar suggestions (in addition to JOSS review) (#227)
* Update index.md

* Update papers.md

* Update faq.md

* Update about.md

* Update bnn-architecture.md

* Update bnn-optimization.md

* Update key-concepts.md

* Update index.md

* Update about.md
