### 2019-09-18 0297f92
Changed FlipRatio aggregation to ONLY_FIRST_REPLICA
