### 2019-09-19 f64e463
Change LarqMetric add_weight default aggregation mode to SUM (like in keras Metric)
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
