### 2019-09-19 e6342fe
Removed add_weight function from LarqMetric as it equalled the parent implementation.
