### 2019-09-19 db603fc
Created LarqMetric class, added register_alias decorator (#225)
* Created LarqMetric class, added register_alias decorator to utils

* added boolean is_weight_metric to FlipRatio so we can distinguish weight and activation metrics

* Added default get_config() and reset_state() functions to _CompatibilityMetric

* Changed FlipRatio aggregation to ONLY_FIRST_REPLICA
