### 2019-09-19 bffe4a4
Add option for an unclipped STE
