+++ larq/quantizers_test.py
@@ -132,6 +132,23 @@ def test_ternarization_with_ternary_weight_networks():
 @pytest.mark.skipif(not tf.executing_eagerly(), reason="requires eager execution")
 def test_ste_grad(fn):
     @np.vectorize
+    def constant_grad(x):
+        return 1.0
+
+    x = np.random.uniform(-5, 5, (8, 3, 3, 16))
+    tf_x = tf.Variable(x)
+    with tf.GradientTape() as tape:
+        activation = fn(tf_x, clip_value=None)
+    grad = tape.gradient(activation, tf_x)
+    np.testing.assert_allclose(grad.numpy(), constant_grad(x))
+
+
+@pytest.mark.parametrize(
+    "fn", [lq.quantizers.ste_sign, lq.quantizers.ste_tern, lq.quantizers.ste_heaviside]
+)
+@pytest.mark.skipif(not tf.executing_eagerly(), reason="requires eager execution")
+def test_identity_ste_grad(fn):
+    @np.vectorize
     def ste_grad(x):
         if np.abs(x) <= 1:
             return 1.0
