### 2019-09-19 625472c
Add option to use unclipped identity STE in quantizers (#232)
* Add option for an unclipped STE

* Added unit tests

* Add Docstring note
