+++ larq/quantizers.py
@@ -146,7 +146,7 @@ def ste_sign(x, clip_value=1.0):
 
     # Arguments
     x: Input tensor.
-    clip_value: Threshold for clipping gradients.
+    clip_value: Threshold for clipping gradients. If `None` gradients are not clipped.
 
     # Returns
     Binarized tensor.
@@ -156,7 +156,8 @@ def ste_sign(x, clip_value=1.0):
       Activations Constrained to +1 or -1](http://arxiv.org/abs/1602.02830)
     """
 
-    x = tf.clip_by_value(x, -clip_value, clip_value)
+    if clip_value is not None:
+        x = tf.clip_by_value(x, -clip_value, clip_value)
 
     return _binarize_with_identity_grad(x)
 
@@ -185,7 +186,7 @@ class SteSign(QuantizerFunctionWrapper):
     ```
 
     # Arguments
-    clip_value: Threshold for clipping gradients.
+    clip_value: Threshold for clipping gradients. If `None` gradients are not clipped.
 
     # References
     - [Binarized Neural Networks: Training Deep Neural Networks with Weights and
@@ -212,7 +213,7 @@ def magnitude_aware_sign(x, clip_value=1.0):
 
     # Arguments
     x: Input tensor
-    clip_value: Threshold for clipping gradients.
+    clip_value: Threshold for clipping gradients. If `None` gradients are not clipped.
 
     # Returns
     Scaled binarized tensor (with values in \\(\\{-a, a\\}\\), where $a$ is a float).
@@ -237,7 +238,7 @@ class MagnitudeAwareSign(QuantizerFunctionWrapper):
     ```
 
     # Arguments
-    clip_value: Threshold for clipping gradients.
+    clip_value: Threshold for clipping gradients. If `None` gradients are not clipped.
 
     # References
     - [Bi-Real Net: Enhancing the Performance of 1-bit CNNs With Improved
@@ -383,7 +384,7 @@ def ste_tern(x, threshold_value=0.05, ternary_weight_networks=False, clip_value=
     threshold_value: The value for the threshold, $\Delta$.
     ternary_weight_networks: Boolean of whether to use the
         Ternary Weight Networks threshold calculation.
-    clip_value: Threshold for clipping gradients.
+    clip_value: Threshold for clipping gradients. If `None` gradients are not clipped.
 
     # Returns
     Ternarized tensor.
@@ -391,7 +392,8 @@ def ste_tern(x, threshold_value=0.05, ternary_weight_networks=False, clip_value=
     # References
     - [Ternary Weight Networks](http://arxiv.org/abs/1605.04711)
     """
-    x = tf.clip_by_value(x, -clip_value, clip_value)
+    if clip_value is not None:
+        x = tf.clip_by_value(x, -clip_value, clip_value)
 
     if ternary_weight_networks:
         threshold = 0.7 * tf.reduce_sum(tf.abs(x)) / tf.cast(tf.size(x), x.dtype)
@@ -441,7 +443,7 @@ class SteTern(QuantizerFunctionWrapper):
     threshold_value: The value for the threshold, $\Delta$.
     ternary_weight_networks: Boolean of whether to use the
         Ternary Weight Networks threshold calculation.
-    clip_value: Threshold for clipping gradients.
+    clip_value: Threshold for clipping gradients. If `None` gradients are not clipped.
 
     # References
     - [Ternary Weight Networks](http://arxiv.org/abs/1605.04711)
@@ -483,7 +485,7 @@ class SteHeaviside(QuantizerFunctionWrapper):
     ```
 
     # Arguments
-    clip_value: Threshold for clipping gradients.
[…]
