### 2019-09-19 ebcc673
Run mkdocs a second time to make sure generated plots are copied
