### 2019-09-19 d4da618
Simplify quantizer constant gradient test (#234)
