+++ larq/quantizers_test.py
@@ -131,16 +131,12 @@ def test_ternarization_with_ternary_weight_networks():
 )
 @pytest.mark.skipif(not tf.executing_eagerly(), reason="requires eager execution")
 def test_identity_ste_grad(fn):
-    @np.vectorize
-    def constant_grad(x):
-        return 1.0
-
     x = np.random.uniform(-5, 5, (8, 3, 3, 16))
     tf_x = tf.Variable(x)
     with tf.GradientTape() as tape:
         activation = fn(tf_x, clip_value=None)
     grad = tape.gradient(activation, tf_x)
-    np.testing.assert_allclose(grad.numpy(), constant_grad(x))
+    np.testing.assert_allclose(grad.numpy(), np.ones_like(x))
 
 
 @pytest.mark.parametrize(
@@ -195,7 +191,6 @@ def test_approx_sign_grad():
 
 @pytest.mark.skipif(not tf.executing_eagerly(), reason="requires eager execution")
 def test_magnitude_aware_sign():
-
     a = np.random.uniform(-2, 2, (3, 2, 2, 3))
     x = tf.Variable(a)
     with tf.GradientTape() as tape:
