### 2019-09-19 9abd535
Switch Larq Zoo examples to a runnable notebook
This changes the Larq Zoo examples to a runnable notebook that also adds better descriptions instead of just adapting the Keras code snippets.

See https://github.com/openjournals/joss-reviews/issues/1746#issuecomment-532738850
