### 2019-09-19 d6c5436
Use Altair/Vega to plot activations and quantizers in API docs (#231)
* Use Altair/Vega to plot activations and quantizers in API docs

This switches the activation and quantizer plots in the API docs from matplotlib to Altair/Vega to make them consistent with the rest of the plots in our docs.

This will also allow us in the future to add interactive tooltips to the plot if we need to.

* Run mkdocs a second time to make sure generated plots are copied
