+++ mkdocs.yml
@@ -66,7 +66,7 @@ markdown_extensions:
       custom_fences:
         - name: plot-activation
           class: larq-activation
-          format: !!python/name:plot_activation.html_format
+          format: !!python/name:plot_altair.plot_activation
         - name: plot-altair
           class: larq-altair
           format: !!python/name:plot_altair.html_format
