### 2019-09-19 f1bb12a
Switch Larq Zoo examples to a runnable notebook (#235)
This changes the Larq Zoo examples to a runnable notebook that also adds better descriptions instead of just adapting the Keras code snippets.

See https://github.com/openjournals/joss-reviews/issues/1746#issuecomment-532738850
