### 2019-09-19 cc8f4b8
Add missing binder dependencies
This makes sure that https://larq.dev/models/examples/ can be executed properly on [Binder](https://mybinder.org)
