### 2019-09-20 93d0f38
Change "Why Larq?" menu item to "About"
Having a FAQ under "Why Larq?" is not really intuitive. I think changing the heading to "About" makes sense or should we name it differently?
