+++ mkdocs.yml
@@ -32,7 +32,7 @@ nav:
       - Papers using Larq: papers.md
       - Contributing Guide: contributing.md
       - Code of Conduct: code-of-conduct.md
-  - Why Larq?:
+  - About:
       - About: about.md
       - FAQ: faq.md
 
