### 2019-09-20 1557713
Clarify installation instructions
This should help people to get started who are not familiar with the TensorFlow ecosystem.
