### 2019-09-20 4e6344f
Add note on how to reach out to us (#237)
This should make it clear how to ask questions or raise issues.
