### 2019-09-20 bfffb2d
Add link to tf.math.sign in lq.math.sign (#238)
This should clarify how `lq.math.sign` differs from `tf.math.sign`.
