### 2019-09-20 551b268
Add link to larq/zoo repo (#242)
This should clarify the location of pretrained models and moves the installation instructions further to the top, to make the distinktion between larq zoo and larq clearer.
