### 2019-09-20 8b60825
Clarify installation instructions (#240)
This should help people to get started who are not familiar with the TensorFlow ecosystem.
