### 2019-09-20 b400e37
Mention students in about page (#241)
