+++ docs/about.md
@@ -2,6 +2,6 @@ Larq is an open-source Binarized Neural Network (BNN) training library on top of
 
 BNNs are the key that will unlock this future. These networks have a much smaller memory footprint than ordinary neural networks, and their bitwise nature makes it possible to execute the computations extremely efficiently.
 
-Larq has been created to make it easier for researchers and developers to explore the possibilities of BNNs. We strongly believe an open approach to knowledge is the way forward, which is why Larq lives under a permissive open-source software [Apache 2.0 licence](https://github.com/larq/larq/blob/master/LICENSE).
+Larq has been created to make it easier for researchers and developers to explore the possibilities of BNNs. We strongly believe an open approach to knowledge is the way forward, which is why Larq lives under a permissive open-source software [Apache 2.0 licence](https://github.com/larq/larq/blob/master/LICENSE). Furthermore, Larq makes it easier for beginners and students to get started with BNNs.
 
 [<img src="https://plumerai.com/assets/img/logo.svg" alt="Plumerai Logo" width="200"/ style="display:block;margin-left:auto;margin-right:auto;margin-top:100px;margin-bottom:100px">](https://plumerai.com)
