### 2019-09-20 d7faae2
Rename "About" --> "Why Larq?"
