### 2019-09-20 fde7266
Remove redundant sentence from about page
