### 2019-09-20 70aaf5f
Added GradientFlow metric
