+++ larq/metrics_test.py
@@ -2,6 +2,7 @@ import tensorflow as tf
 from larq import metrics
 from tensorflow.python.keras import keras_parameterized
 import pytest
+import numpy as np
 
 
 def test_scope():
@@ -81,3 +82,126 @@ class FlipRatioTest(keras_parameterized.TestCase):
         self.assertAllClose(1.5, total)
         self.assertAllClose(3, count)
         self.assertAllClose(1.5 / 2, result)
+
+
+class GradientFlowTest(keras_parameterized.TestCase):
+    def test_config(self):
+        mcv = metrics.GradientFlow(name="mcv", threshold=5.0, dtype="float32")
+        self.assertEqual(mcv.name, "mcv")
+        self.assertTrue(mcv.stateful)
+        self.assertEqual(mcv.dtype, tf.float32)
+
+        mcv2 = metrics.GradientFlow.from_config(mcv.get_config())
+        self.assertEqual(mcv2.name, "mcv")
+        self.assertTrue(mcv2.stateful)
+        self.assertEqual(mcv2.dtype, tf.float32)
+        self.assertAllClose(mcv2.threshold, 5.0)
+
+    def test_metric(self):
+        mcv = metrics.GradientFlow(threshold=1.0, dtype="float32")
+        self.evaluate(tf.compat.v1.variables_initializer(mcv.variables))
+
+        self.assertAllClose(0.0, mcv.result())
+        self.assertAllClose(0, self.evaluate(mcv.total_value))
+
+        # 2/4 all pixels in the batch receive a gradient
+        self.evaluate(mcv.update_state([[-2, 0.5], [-1.1, 0.9]]))
+        self.assertAllClose(0.5, self.evaluate(mcv.total_value))
+        self.assertAllClose(0.5, mcv.result())
+
+        # No gradients
+        self.evaluate(mcv.update_state([[-5, 8], [2, 1.03]]))
+        self.assertAllClose(0.5, self.evaluate(mcv.total_value))
+        self.assertAllClose(mcv.result(), 0.25)
+
+        # batch size 2, and 4 feature maps of 4x4 pixels
+        # First sample: 1/64 pixels receives a gradient. Second sample: None.
+        state = np.ones((2, 4, 4, 4)).astype(np.float32) + 0.1
+        state[0, 0, 0, 0] = -1  # -1 does not count as saturation. -1.0001 does.
+        self.evaluate(mcv.update_state(state))
+        self.assertAllClose((1.0 / 128.0 + 0.5), self.evaluate(mcv.total_value))
+        self.assertAllClose(np.mean([0.5, 0.0, 1.0 / 128.0]), mcv.result())
+
+    @pytest.mark.skipif(tf.executing_eagerly(), reason="only applies to graph mode")
+    def test_metric_in_graph_mode(self):
+        mcv = metrics.GradientFlow(threshold=1.0)
+
+        new_state = tf.compat.v1.placeholder(dtype=tf.float32, shape=[2, 2, 2, 3])
+        update_state_op = mcv.update_state(new_state)
+        metric_value = mcv.result()
+
+        with tf.compat.v1.Session() as sess:
+            sess.run(tf.compat.v1.variables_initializer(mcv.variables))
+
+            # Batch 1: gradient flow 15/24
+            sess.run(
+                update_state_op,
+                feed_dict={
+                    # Formatted as BxCxWxH now, but the swapaxes ensures that we get BxWxHxC.
+                    new_state: np.array(
+                        [
+                            # Sample 1: 7/12 pixels receive gradient.
+                            [
+                                # Channel 1
+                                [[0.5, -0.5], [-3.0, -0.9]],
+                                # Channel 2
+                                [[0, -0.4], [-2, 5]],
+                                # Channel 3
+                                [[0, -0.4], [-2, 5]],
[…]
