### 2019-09-20 ce44818
Revert "ran black code formatter"
This reverts commit f8d72180bca4b18381b1d68c65720f26a6eb30f5.
