### 2019-09-20 7cd285b
Revert "Added GradientFlow metric"
This reverts commit 70aaf5f51ab4aa3100c405d14c80ecc92f66d1de.
