### 2019-09-20 bb9101c
Change "Why Larq?" menu item to "About" (#239)
* Change "Why Larq?" menu item to "About"

Having a FAQ under "Why Larq?" is not really intuitive. I think changing the heading to "About" makes sense or should we name it differently?

* Rename "About" --> "Why Larq?"
