+++ mkdocs.yml
@@ -32,8 +32,8 @@ nav:
       - Papers using Larq: papers.md
       - Contributing Guide: contributing.md
       - Code of Conduct: code-of-conduct.md
-  - Why Larq?:
-      - About: about.md
+  - About:
+      - Why Larq?: about.md
       - FAQ: faq.md
 
 repo_url: https://github.com/larq/larq
