### 2019-09-20 9611d7a
Added numpy import to metric_test, removed indendation in GradientFlow docstring
