### 2019-09-20 9935440
Integrated code cleanup suggestions by Lukas
