### 2019-09-23 1dfefca
Bump gast from 0.2.2 to 0.3.2
Bumps [gast](https://github.com/serge-sans-paille/gast) from 0.2.2 to 0.3.2.
- [Release notes](https://github.com/serge-sans-paille/gast/releases)
- [Commits](https://github.com/serge-sans-paille/gast/commits)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
