### 2019-09-23 b842213
Bump mknotebooks from 0.1.6 to 0.1.7
Bumps [mknotebooks](https://github.com/greenape/mknotebooks) from 0.1.6 to 0.1.7.
- [Release notes](https://github.com/greenape/mknotebooks/releases)
- [Commits](https://github.com/greenape/mknotebooks/compare/0.1.6...0.1.7)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
