+++ setup.py
@@ -36,7 +36,7 @@ setup(
             "mkdocs==1.0.4",
             "mkdocs-material==4.4.2",
             "pymdown-extensions==6.1",
-            "mknotebooks==0.1.6",
+            "mknotebooks==0.1.7",
             "mkdocs-minify-plugin==0.2.1",
             "larq-zoo==0.4.1",
             "altair==3.2.0",
