### 2019-09-24 186565c
Properly format link to paper in markdown
Co-Authored-By: Koen Helwegen <koen@plumerai.com>
