+++ azure-pipelines.yml
@@ -19,10 +19,10 @@ jobs:
           pytest.args: "--cov=larq --cov-report=html --cov-config=.coveragerc"
         Python37_TF115:
           python.version: "3.7"
-          tensorflow: "tensorflow-cpu==1.15.0-rc0"
+          tensorflow: "tensorflow-cpu==1.15.0-rc1"
         Python37_TF2:
           python.version: "3.7"
-          tensorflow: "tensorflow==2.0.0-rc1"
+          tensorflow: "tensorflow==2.0.0-rc2"
 
     steps:
       - task: UsePythonVersion@0
