+++ docs/guides/key-concepts.md
@@ -1,45 +1,42 @@
-If you are new to Larq and/or Binarized Neural Networks (BNNs), this is the right place to start. Below we summarize the key concepts you need to understand to work with BNNs.
+If you are new to Larq and/or Binarized Neural Networks (BNNs), this is the right place to start. 
+Below, we summarize the key concepts you need to understand to work with BNNs.
 
 ## Quantizer
 
-A [quantizer](https://larq.dev/api/quantizers/) defines an operation that quantizes a vector, as well as a pseudo-gradient that is used for automatic differentiation. This pseudo-gradient is in general not the true gradient.
+The transformation from high-precision Neural Networks to Quantized Neural Networks (QNNs) is achieved by [quantization](https://en.wikipedia.org/wiki/Quantization_(signal_processing)).
+This is the process of mapping a large set of, often continuous, values to a smaller countable set. 
+Binarized Neural Networks are a special case of QNNs, where the quantization output is binary.
 
-Generally, you will find quantizers throughout the network to quantize activations. This is because most layers output integers, even if all inputs are binary, because they sum over multiple binary values.
+In `larq`, A [quantizer](https://larq.dev/api/quantizers/) defines the way of transforming a full precision input to a quantized output and the pseudo-gradient method used for the backwards pass.
+The latter is called pseudo-gradient, as it is in general not the true gradient.
 
-It is also common to apply quantizers to the weights during training. This is necessary when relying on real-valued latent-weights to accumulate non-binary update steps, a common optimization strategy for BNNs. After training is finished, the real-valued weights and associated quantization operations can be discarded.
+Generally, you will find quantizers throughout the network to quantize activations. 
+This is because most layers output integers, even if all inputs are binary, because they sum over multiple binary values.
+
+It is also common to apply quantizers to the weights during training. 
+This is necessary when relying on real-valued latent-weights to accumulate non-binary update steps, a common optimization strategy for BNNs. 
+After training is finished, the real-valued weights and associated quantization operations can be discarded.
 
 ### Pseudo-Gradient
 
-The true gradient of a quantizer is in general zero almost everywhere and therefore cannot be used for gradient descent. Instead, the optimization of BNNs relies on what we call pseudo-gradients, which are used during back-propagation. In the documentation for each quantizer you will find the definition and a graph of the pseudo-gradient.
-
-### Using Quantizers as Activations
-
-Although quantizers are usually passed to specialized arguments (see [Quantized Layers](#quantized-layers)), they can be used just like `tf.keras` activations:
-
-```python
-# Use a quantizer als activation
-y = larq.layers.QuantDense(512, activation="ste_sign")(x)
-```
-
-Just like activations, quantizers can also be used on their own:
-
-```python
-# The two lines below are equivalent
-x_binarized = larq.quantizers.ste_sign(x)
-x_binarized = tf.keras.layers.Activation("ste_sign")(x)
-```
+The true gradient of a quantizer is in general zero almost everywhere and therefore cannot be used for gradient descent. 
+Instead, the optimization of BNNs relies on what we call pseudo-gradients, which are used during back-propagation. 
+In the documentation for each quantizer you will find the definition and a graph of the pseudo-gradient.
 
 ## Quantized Layers
 
[…]
