### 2019-09-26 #1746 @sbrugman
Hi @lgeiger,

We can see great progress at larq. Keep it up.

This round of feedback complements the earlier comments.

### Software paper

- *Authors*: We need to discuss what we think of "PlumerAI team" as an author, as anonymous authorship is unconventional. @terrytangyuan, do you know of any precedent of this at the JOSS?
I would propose to make a distinction between authors and auxiliaries. Authors are mentioned in the paper and auxiliaries can be acknowledged at appropriate places, such as an acknowledgement section and/or in the documentation.

- *Research applications*: What would the ease of use concretely mean? Are the currently implemented specialized optimizers, training metrics and profiling tools providing the user with enough benefit of using larq? You could demonstrate this by for example going through the steps to implement a recent paper in larq ([this paper might be a good candidate, source code in appendix](https://arxiv.org/ftp/arxiv/papers/1905/1905.11452.pdf)). This point is critical as without research clear application, the added value is primarily educational.
In that case, you could consider JOSE as @casperdcl mentions.

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

- *Paper:* The paper is around 400 words (target is 250-1000).
There are four sections: introduction, package description, background and target audience/future work/goals. State of the field is missing. I suggest taking more advantage of the space you have left and make crystal clear why and how your target audience can use larq.

Good luck. I am looking forward to the next revision of the draft.
