### 2019-09-27 15a49da
Update docs/guides/key-concepts.md
Co-Authored-By: Koen Helwegen <kg.helwegen@gmail.com>
