### 2019-09-27 f82f6ca
Update docs/guides/key-concepts.md
Co-Authored-By: Koen Helwegen <kg.helwegen@gmail.com>
