### 2019-09-28 #1746 @terrytangyuan
Regarding the authorship, I think we would need explicit author names. If this is not possible to obtain, I’d be curious to see why it’s so.
