### 2019-09-30 fce0ee1
Bump larq-zoo from 0.4.1 to 0.4.2
