+++ setup.py
@@ -38,7 +38,7 @@ setup(
             "pymdown-extensions==6.1",
             "mknotebooks==0.1.7",
             "mkdocs-minify-plugin==0.2.1",
-            "larq-zoo==0.4.1",
+            "larq-zoo==0.4.2",
             "altair==3.2.0",
             "pandas==0.25.1",
         ],
