### 2019-09-30 876d3e9
DoReFa-Net (#253)
* Add plot for dorefa_net

* Add DoReFa to index

* Bump larq-zoo from 0.4.1 to 0.4.2

* Update apidocs.yml
