### 2019-09-30 159ddaf
Add formula for swish sign
