### 2019-09-30 555892e
Reference formula for magnitude aware sign
