### 2019-09-30 d44620f
Add formula for swish sign and improve docs for magnitude aware sign (#257)
* Add formula for swish sign

* Reference formula for magnitude aware sign
