### 2019-09-30 3e9a39f
Cite BNN NeurIPS paper
This adds the citation of the original NeurIPS paper to our website and the JOSS paper
