### 2019-09-30 2722b5b
Mathematically clarify binarization in guide
This should make it a bit clearer what happens during quantization of a BNN.
