+++ docs/guides/key-concepts.md
@@ -1,31 +1,34 @@
-If you are new to Larq and/or Binarized Neural Networks (BNNs), this is the right place to start. 
+If you are new to Larq and/or Binarized Neural Networks (BNNs), this is the right place to start.
 Below, we summarize the key concepts you need to understand to work with BNNs.
 
 ## Quantizer
 
 The transformation from high-precision Neural Networks to Quantized Neural Networks (QNNs) is achieved by [quantization](https://en.wikipedia.org/wiki/Quantization_(signal_processing)).
-This is the process of mapping a large set of, often continuous, values to a smaller countable set. 
-Binarized Neural Networks are a special case of QNNs, where the quantization output is binary.
+This is the process of mapping a large set of, often continuous, values to a smaller countable set.
+Binarized Neural Networks are a special case of QNNs, where the quantization output \(x_q\) is binary:
+\\[
+x_q = q(x), \quad x_q \in \\{-1, +1\\}, x \in \mathbb{R}
+\\]
 
-In `larq`, A [quantizer](https://larq.dev/api/quantizers/) defines the way of transforming a full precision input to a quantized output and the pseudo-gradient method used for the backwards pass.
+In `larq`, A [quantizer](https://larq.dev/api/quantizers/) \(q\) defines the way of transforming a full precision input to a quantized output and the pseudo-gradient method used for the backwards pass.
 The latter is called pseudo-gradient, as it is in general not the true gradient.
 
-Generally, you will find quantizers throughout the network to quantize activations. 
+Generally, you will find quantizers throughout the network to quantize activations.
 This is because most layers output integers, even if all inputs are binary, because they sum over multiple binary values.
 
-It is also common to apply quantizers to the weights during training. 
-This is necessary when relying on real-valued latent-weights to accumulate non-binary update steps, a common optimization strategy for BNNs. 
+It is also common to apply quantizers to the weights during training.
+This is necessary when relying on real-valued latent-weights to accumulate non-binary update steps, a common optimization strategy for BNNs.
 After training is finished, the real-valued weights and associated quantization operations can be discarded.
 
 ### Pseudo-Gradient
 
-The true gradient of a quantizer is in general zero almost everywhere and therefore cannot be used for gradient descent. 
-Instead, the optimization of BNNs relies on what we call pseudo-gradients, which are used during back-propagation. 
+The true gradient of a quantizer is in general zero almost everywhere and therefore cannot be used for gradient descent.
+Instead, the optimization of BNNs relies on what we call pseudo-gradients, which are used during back-propagation.
 In the documentation for each quantizer you will find the definition and a graph of the pseudo-gradient.
 
 ## Quantized Layers
 
-Each [quantized layer](https://larq.dev/api/layers/) accepts an `input_quantizer` and a `kernel_quantizer` that describe the way of quantizing the incoming activations and weights of the layer respectively. 
+Each [quantized layer](https://larq.dev/api/layers/) accepts an `input_quantizer` and a `kernel_quantizer` that describe the way of quantizing the incoming activations and weights of the layer respectively.
 If both `input_quantizer` and `kernel_quantizer` are `None` the layer is equivalent to a full precision layer.
 
 A quantized layer computes activations \(\boldsymbol{y}\) as:
@@ -34,7 +37,7 @@ A quantized layer computes activations \(\boldsymbol{y}\) as:
 \boldsymbol{y} = \sigma(f(q_{\, \mathrm{kernel}}(\boldsymbol{w}), q_{\, \mathrm{input}}(\boldsymbol{x})) + b)
 \\]
 
[…]
