### 2019-09-30 ead3049
Cite BNN NeurIPS paper (#258)
* Cite BNN NeurIPS paper

This adds the citation of the original NeurIPS paper to our website and the JOSS paper
