+++ paper/paper.bib
@@ -27,16 +27,13 @@
     adsurl       = {https://ui.adsabs.harvard.edu/abs/2019arXiv190602107H},
     adsnote      = {Provided by the SAO/NASA Astrophysics Data System}
 }
-@article{Courbariaux2016,
-    title        = {Binarized Neural Networks: Training Deep Neural Networks with Weights and Activations Constrained to +1 or -1},
-    author       = {Courbariaux, Matthieu and Hubara, Itay and Soudry, Daniel and El-Yaniv, Ran and Bengio, Yoshua},
-    year         = 2016,
-    doi          = {10.1109/CVPR.2016.90},
-    isbn         = 9781510829008,
-    issn         = {1664-1078},
-    url          = {http://arxiv.org/abs/1602.02830},
-    archiveprefix = {arXiv},
-    arxivid      = {1602.02830},
-    eprint       = {1602.02830},
-    pmid         = 23554596
+@incollection{NIPS2016_6573,
+    title        = {Binarized Neural Networks},
+    author       = {Hubara, Itay and Courbariaux, Matthieu and Soudry, Daniel and El-Yaniv, Ran and Bengio, Yoshua},
+    booktitle    = {Advances in Neural Information Processing Systems 29},
+    editor       = {D. D. Lee and M. Sugiyama and U. V. Luxburg and I. Guyon and R. Garnett},
+    pages        = {4107--4115},
+    year         = {2016},
+    publisher    = {Curran Associates, Inc.},
+    url          = {http://papers.nips.cc/paper/6573-binarized-neural-networks.pdf}
 }
