### 2019-09-30 448ac0b
Extend example on front page and remove it from user guide
