+++ docs/guides/key-concepts.md
@@ -103,58 +103,6 @@ x_binarized = larq.quantizers.ste_sign(x)
 x_binarized = tf.keras.layers.Activation("ste_sign")(x)
 ```
 
-### Defining a simple Binarized Neural Network
-A simple fully-connected Binarized Neural Network (BNN) using the [Straight-Through Estimator](https://larq.dev/api/quantizers/#ste_sign) can be defined in just a few lines of code using either the Keras sequential, functional or model subclassing APIs:
-
-```python tab="Larq 1-bit model"
-model = tf.keras.models.Sequential([
-    tf.keras.layers.Flatten(),
-    larq.layers.QuantDense(512,
-                           kernel_quantizer="ste_sign",
-                           kernel_constraint="weight_clip"),
-    larq.layers.QuantDense(10,
-                           input_quantizer="ste_sign",
-                           kernel_quantizer="ste_sign",
-                           kernel_constraint="weight_clip",
-                           activation="softmax")])
-```
-
-```python tab="Larq 1-bit model functional"
-x = tf.keras.Input(shape=(28, 28, 1))
-y = tf.keras.layers.Flatten()(x)
-y = larq.layers.QuantDense(512,
-                           kernel_quantizer="ste_sign",
-                           kernel_constraint="weight_clip")(y)
-y = larq.layers.QuantDense(10,
-                           input_quantizer="ste_sign",
-                           kernel_quantizer="ste_sign",
-                           kernel_constraint="weight_clip",
-                           activation="softmax")(y)
-model = tf.keras.Model(inputs=x, outputs=y)
-```
-
-```python tab="Larq 1-bit model subclassing"
-class MyModel(tf.keras.Model):
-    def __init__(self):
-        super().__init__()
-        self.flatten = tf.keras.layers.Flatten()
-        self.dense1 = larq.layers.QuantDense(512,
-                                             kernel_quantizer="ste_sign",
-                                             kernel_constraint="weight_clip")
-        self.dense2 = larq.layers.QuantDense(10,
-                                             input_quantizer="ste_sign",
-                                             kernel_quantizer="ste_sign",
-                                             kernel_constraint="weight_clip",
-                                             activation="softmax")
-
-    def call(self, inputs):
-        x = self.flatten(inputs)
-        x = self.dense1(x)
-        return self.dense2(x)
-
-model = MyModel()
-```
-
 ## Using Custom Quantizers
 
 Quantizers are functions that transform a full-precision input to a quantized output.
