+++ docs/index.md
@@ -6,11 +6,15 @@ Existing deep neural networks use 32 bits, 16 bits or 8 bits to encode each weig
 
 ## Getting Started
 
-To build a QNN, Larq introduces the concept of [quantized layers](https://larq.dev/api/layers/) and [quantizers](https://larq.dev/api/quantizers/). A quantizer defines the way of transforming a full precision input to a quantized output and the pseudo-gradient method used for the backwards pass. Each quantized layer requires an `input_quantizer` and a `kernel_quantizer` that describe the way of quantizing the incoming activations and weights of the layer respectively. If both `input_quantizer` and `kernel_quantizer` are `None` the layer is equivalent to a full precision layer.
+To build a QNN, Larq introduces the concept of [quantized layers](https://larq.dev/api/layers/) and [quantizers](https://larq.dev/api/quantizers/). A quantizer defines the way of transforming a full precision input to a quantized output and the pseudo-gradient method used for the backwards pass. Each quantized layer requires an `input_quantizer` and a `kernel_quantizer` that describe the way of quantizing the incoming activations and weights of the layer respectively. If both `input_quantizer` and `kernel_quantizer` are `None` the layer is equivalent to a full precision layer. This layer can be used inside a [Keras model](https://www.tensorflow.org/alpha/guide/keras/overview#sequential_model) or with a [custom training loop](https://www.tensorflow.org/alpha/guide/keras/overview#model_subclassing).
 
-You can define a simple binarized fully-connected Keras model using the [Straight-Through Estimator](https://larq.dev/api/quantizers/#ste_sign) the following way:
+For a detailed explanation checkout our [user guide](https://larq.dev/guides/key-concepts/).
 
-```python
+### Defining a simple BNN
+
+A simple fully-connected BNN using the [Straight-Through Estimator](https://larq.dev/api/quantizers/#ste_sign) can be defined in just a few lines of code using either the Keras sequential, functional or model subclassing APIs:
+
+```python tab="Larq 1-bit model"
 model = tf.keras.models.Sequential([
     tf.keras.layers.Flatten(),
     larq.layers.QuantDense(512,
@@ -23,7 +27,41 @@ model = tf.keras.models.Sequential([
                            activation="softmax")])
 ```
 
-This layer can be used inside a [Keras model](https://www.tensorflow.org/alpha/guide/keras/overview#sequential_model) or with a [custom training loop](https://www.tensorflow.org/alpha/guide/keras/overview#model_subclassing).
+```python tab="Larq 1-bit model functional"
+x = tf.keras.Input(shape=(28, 28, 1))
+y = tf.keras.layers.Flatten()(x)
+y = larq.layers.QuantDense(512,
+                           kernel_quantizer="ste_sign",
+                           kernel_constraint="weight_clip")(y)
+y = larq.layers.QuantDense(10,
+                           input_quantizer="ste_sign",
+                           kernel_quantizer="ste_sign",
+                           kernel_constraint="weight_clip",
+                           activation="softmax")(y)
+model = tf.keras.Model(inputs=x, outputs=y)
+```
+
+```python tab="Larq 1-bit model subclassing"
+class MyModel(tf.keras.Model):
+    def __init__(self):
+        super().__init__()
+        self.flatten = tf.keras.layers.Flatten()
+        self.dense1 = larq.layers.QuantDense(512,
+                                             kernel_quantizer="ste_sign",
+                                             kernel_constraint="weight_clip")
+        self.dense2 = larq.layers.QuantDense(10,
+                                             input_quantizer="ste_sign",
+                                             kernel_quantizer="ste_sign",
+                                             kernel_constraint="weight_clip",
[…]
