### 2019-09-30 7a24467
Improve summary and clarify target audience
