### 2019-09-30 larq/larq#262 @lgeiger
Add section about performance claims of quantized/binarized NNs
In https://github.com/openjournals/joss-reviews/issues/1746 @sbrugman raised a very valid point that is currently missing in our guides and documentation:

> The documentation and paper make no claims on the comparison of performance between Quantized Neural Networks and their high-precision weight and bias counterparts. At the time of writing, there is clearly a trade-off (although the authors are working hard to close the gap). I suggest being transparent in this difference and possibly opening the option of educating users on when a few percents of accuracy on a specific dataset are essential and when not. This might be especially relevant for people working in the industry.

I think it would be good to add a small paragraph to our FAQ or the guides that briefly explains the story about efficiency and the challenges and of quantized networks as well as what theoretical speedups someone can expect. In general this always depends on the choice of networks and hardware used to run the models, but I agree that we should be transparent about this.

@jamescook106 @koenhelwegen Do we have some resources we could use to elaborate on our claims on the home page?
