### 2019-09-30 f550a35
Extend example on front page and remove it from user guide (#261)
* Extend example on front page and remove it from user guide

* Improve summary and clarify target audience
