### 2019-09-30 4242c97
Bump larq-zoo from 0.4.1 to 0.4.2
Bumps [larq-zoo](https://github.com/plumerai/larq-zoo) from 0.4.1 to 0.4.2.
- [Release notes](https://github.com/plumerai/larq-zoo/releases)
- [Commits](https://github.com/plumerai/larq-zoo/compare/v0.4.1...v0.4.2)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
