+++ requirements.txt
@@ -1,6 +1,6 @@
 tensorflow==1.14.0
 tensorflow_datasets==1.2.0
-larq_zoo==0.4.1
+larq_zoo==0.4.2
 Pillow==6.1.0
 matplotlib==3.1.1
 gast==0.2.2  # See https://github.com/tensorflow/tensorflow/issues/32319
