### 2019-10-01 55d1ec6
Extend site description (#260)
This hopefully helps us with visability on google search
