### 2019-10-02 61ae38e
Explicitely clip gradients of quantizers in backwards pass
