### 2019-10-02 fa5ef93
Fix use of tf.custom_gradient since it doesn't support keyword arguments
