### 2019-10-02 f8c346b
Explicitely clip gradients of quantizers in backwards pass (#267)
* Explicitely clip gradients of quantizers in backwards pass

* Fix use of tf.custom_gradient since it doesn't support keyword arguments
