+++ .github/pr_labels.yml
@@ -1,0 +1,10 @@
+version: "1"
+invalidStatus: "pending"
+labelRule:
+  values:
+    - "bug"
+    - "documentation"
+    - "internal improvement"
+    - "performance"
+    - "dependencies"
+    - "feature"
