### 2019-10-03 cca4c7f
Move linting to GitHub actions
