+++ .github/workflows/check_and_publish.yml
@@ -1,25 +1,30 @@
-name: Publish Larq
+name: check
 
 on: [push]
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v1
-    - name: Set up Python 3.7
-      uses: actions/setup-python@v1
-      with:
-        python-version: 3.7
-    - name: Upgrade setuptools
-      run: pip install --upgrade setuptools wheel
-    - name: Build distribution
-      run: python setup.py sdist bdist_wheel
-    - name: Publish package
-      if: github.event_name == 'push' && startsWith(github.event.ref, 'refs/tags')
-      uses: pypa/gh-action-pypi-publish@master
-      with:
-        user: __token__
-        password: ${{ secrets.pypi_password }}
+      - uses: actions/checkout@v1
+      - name: Set up Python 3.7
+        uses: actions/setup-python@v1
+        with:
+          python-version: 3.7
+      - name: Upgrade setuptools
+        run: pip install --upgrade setuptools wheel
+      - name: Install Lint Dependencies
+        run: pip install pyflakes==2.1.1 black==19.3b0
+      - name: Run PyFlakes
+        run: pyflakes .
+      - name: Black code style
+        run: black . --check --target-version py36 --exclude 'build/|buck-out/|dzzist/|_build/|\.git/|\.hg/|\.mypy_cache/|\.tox/|\.venv/|larq/snapshots/'
+      - name: Build distribution
+        run: python setup.py sdist bdist_wheel
+      - name: Publish package
+        if: github.event_name == 'push' && startsWith(github.event.ref, 'refs/tags')
+        uses: pypa/gh-action-pypi-publish@master
+        with:
+          user: __token__
+          password: ${{ secrets.pypi_password }}
