### 2019-10-03 fe5c42b
Split Lint and Publish workflow
