### 2019-10-03 d4c5850
Update release guide
