### 2019-10-03 0ac0f3d
Add github action to publish package (#272)
* Add github action to publish package

* Move linting to GitHub actions

* Split Lint and Publish workflow

* Update release guide
