+++ .github/workflows/lint.yml
@@ -1,0 +1,20 @@
+name: Lint
+
+on: [push]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v1
+      - name: Set up Python 3.7
+        uses: actions/setup-python@v1
+        with:
+          python-version: 3.7
+      - name: Install Lint Dependencies
+        run: pip install pyflakes==2.1.1 black==19.3b0
+      - name: Run PyFlakes
+        run: pyflakes .
+      - name: Black code style
+        run: black . --check --target-version py36 --exclude 'build/|buck-out/|dzzist/|_build/|\.git/|\.hg/|\.mypy_cache/|\.tox/|\.venv/|larq/snapshots/'
