+++ azure-pipelines.yml
@@ -38,21 +38,3 @@ jobs:
 
       - script: pytest . -vv $(pytest.args)
         displayName: "pytest"
-
-  - job: "Lint"
-    pool:
-      vmImage: "Ubuntu-16.04"
-    steps:
-      - task: UsePythonVersion@0
-        inputs:
-          versionSpec: "3.7"
-          architecture: "x64"
-
-      - script: pip install pyflakes==2.1.1 black==19.3b0
-        displayName: "Install linting dependencies"
-
-      - script: pyflakes .
-        displayName: "pyflakes"
-
-      - script: black . --check --target-version py36 --exclude 'build/|buck-out/|dzzist/|_build/|\.git/|\.hg/|\.mypy_cache/|\.tox/|\.venv/|larq/snapshots/'
-        displayName: "Black code style"
