+++ .github/release-drafter.yml
@@ -1,0 +1,15 @@
+categories:
+  - title: ":warning: Breaking Changes :warning:"
+    label: "breaking-change"
+  - title: ":tada: Features"
+    label: "feature"
+  - title: ":bug: Bug Fixes"
+    label: "bug"
+  - title: ":book: Documentation"
+    label: "documentation"
+  - title: ":construction_worker_man: Internal Improvements"
+    label: "internal-improvement"
+  - title: ":arrow_up: Depenencies"
+    label: "dependencies"
+exclude-labels:
+  - "skip-changelog"
