### 2019-10-03 bd4145d
Update .github/release-drafter.yml
Co-Authored-By: Adam Hillier <7688302+AdamHillier@users.noreply.github.com>
