### 2019-10-03 9b2a312
Automatically draft release notes (#273)
* Automatically draft release notes

This will automatically draft release notes from merged PRs using https://github.com/toolmantim/release-drafter

* Update .github/release-drafter.yml

Co-Authored-By: Adam Hillier <7688302+AdamHillier@users.noreply.github.com>
