### 2019-10-03 69540f7
Reorder quantizers so they show up grouped in the docs (#274)
This is inspired by #246
