### 2019-10-03 027cb27
Add Release Drafter as GH action
This should automatically draft release notes on once a PR is merged.
