+++ .github/workflows/release-notes.yml
@@ -1,0 +1,11 @@
+name: Release Notes
+
+on: [push]
+
+jobs:
+  update_draft_release:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: toolmantim/release-drafter@v5.2.0
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
