### 2019-10-03 fa7a6e0
Add Release Drafter as GH action (#275)
* Add Release Drafter as GH action

This should automatically draft release notes on once a PR is merged.

* Fix config
