### 2019-10-03 0fa3f91
Only run release notes action on pushed to master
This is hopefully the last PR that's needed to configure this bot correctly. Since the template is now setup correctly we could also use the GitHub app, but I think the GitHub action gives us more control (e.g. if we only want to run it when the tag is pushed).
