### 2019-10-03 ee56400
Run unit tests with GitHub actions
