+++ .github/workflows/unittest.yml
@@ -1,0 +1,41 @@
+name: Unittest
+
+on: [push]
+
+jobs:
+  tests:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        tf-version: [1.13.1, 1.14.0, 1.15.0-rc2, 2.0.0]
+
+    steps:
+      - uses: actions/checkout@v1
+      - name: Set up Python
+        uses: actions/setup-python@v1
+        with:
+          python-version: 3.7
+      - name: Install dependencies
+        run: |
+          pip install tensorflow==${{matrix.tf-version}}
+          pip install -e .[test]
+      - name: Test with pytest
+        run: pytest .
+  coverage:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v1
+      - name: Set up Python
+        uses: actions/setup-python@v1
+        with:
+          python-version: 3.6
+      - name: Install dependencies
+        run: |
+          pip install tensorflow==2.0.0
+          pip install -e .[test]
+      - name: Generate coverage report
+        run: pytest . --cov=larq --cov-report=xml --cov-config=.coveragerc
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v1.0.2
+        with:
+          token: ${{secrets.CODECOV_TOKEN}}
