### 2019-10-03 8ee1612
Don't use codecov action to speed up docker pull
