### 2019-10-03 d951b78
Modify badges
