### 2019-10-03 995f56e
Tune code coverage
