### 2019-10-04 a489479
Run unit tests with GitHub actions (#277)
* Run unit tests with GitHub actions

* Don't use codecov action to speed up docker pull

* Don't comment on PR

* Modify badges

* Remove azure pipelines

* Tune code coverage
