+++ azure-pipelines.yml
@@ -1,40 +1,0 @@
-trigger:
-  - master
-
-pr:
-  - master
-
-jobs:
-  - job: "Test"
-    pool:
-      vmImage: "Ubuntu-16.04"
-    strategy:
-      matrix:
-        Python36_TF113:
-          python.version: "3.6"
-          tensorflow: "tensorflow==1.13.1"
-        Python37_TF114:
-          python.version: "3.7"
-          tensorflow: "tensorflow==1.14.0"
-          pytest.args: "--cov=larq --cov-report=html --cov-config=.coveragerc"
-        Python37_TF115:
-          python.version: "3.7"
-          tensorflow: "tensorflow-cpu==1.15.0-rc1"
-        Python37_TF2:
-          python.version: "3.7"
-          tensorflow: "tensorflow==2.0.0"
-
-    steps:
-      - task: UsePythonVersion@0
-        inputs:
-          versionSpec: "$(python.version)"
-          architecture: "x64"
-
-      - script: |
-          pip install $(tensorflow)
-          pip install -e .[test]
-          pip install pytest-azurepipelines
-        displayName: "Install dependencies"
-
-      - script: pytest . -vv $(pytest.args)
-        displayName: "pytest"
