### 2019-10-04 9ba1799
Bump pytest-cov from 2.7.1 to 2.8.0 (#278)
Bumps [pytest-cov](https://github.com/pytest-dev/pytest-cov) from 2.7.1 to 2.8.0.
- [Release notes](https://github.com/pytest-dev/pytest-cov/releases)
- [Changelog](https://github.com/pytest-dev/pytest-cov/blob/master/CHANGELOG.rst)
- [Commits](https://github.com/pytest-dev/pytest-cov/compare/v2.7.1...v2.8.0)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
