### 2019-10-08 85f4942
Add explanation of scoped metric definition
Based on explanation by @lgeiger
