### 2019-10-09 d669797
:arrow_up: vega-embed@6
This brings us to the newest version of vega-embed used for rendering our interactive plots.
