### 2019-10-10 f02da26
Fixed test bugs and actually test twn
