### 2019-10-10 7f780d5
Adam's Suggestions
Co-Authored-By: Adam Hillier <7688302+AdamHillier@users.noreply.github.com>
