### 2019-10-10 c2d3f9f
Quantizer tests improvements (#284)
* WIP

* WIP

* Fixed test bugs and actually test twn

* Final Initial test rewrite

* Adam's Suggestions

Co-Authored-By: Adam Hillier <7688302+AdamHillier@users.noreply.github.com>
