### 2019-10-14 5659755
Fix iteration bug in bop
