### 2019-10-14 f354288
Add additional optimizer tests
