### 2019-10-14 b642370
Update larq/optimizers_v2.py
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
