+++ larq/optimizers_v2.py
@@ -58,11 +58,12 @@ class Bop(tf.keras.optimizers.Optimizer):
                 self.add_slot(var, "m")
 
     def apply_gradients(self, grads_and_vars, name=None):
-        # convert from zip object so we can loop multiple times:
-        grads_and_vars = list(grads_and_vars)
-
-        bin_grads_and_vars = [(g, v) for g, v in grads_and_vars if self.is_binary(v)]
-        fp_grads_and_vars = [(g, v) for g, v in grads_and_vars if not self.is_binary(v)]
+        bin_grads_and_vars, fp_grads_and_vars = [], []
+        for grad, var in grads_and_vars:
+            if self.is_binary(var):
+                bin_grads_and_vars.append((grad, var))
+            else:
+                fp_grads_and_vars.append((grad, var))
 
         bin_train_op = super().apply_gradients(bin_grads_and_vars, name=name)
 
