### 2019-10-14 8c73806
remove empty variable list fix
