### 2019-10-14 043e4e8
Fix iteration bug in Bop with TensorFlow 2.0.0 (#287)
* Fix iteration bug in bop

* Add additional optimizer tests

* Update larq/optimizers_v2.py

Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>

* single loop

* remove empty variable list fix
