### 2019-10-14 933c176
Fix links on front page
This changes the links to the docs of the stable release and resolves the redirect in the link to the Python website.
