+++ README.md
@@ -26,7 +26,7 @@ model = tf.keras.models.Sequential([
                            activation="softmax")])
 ```
 
-This layer can be used inside a [Keras model](https://www.tensorflow.org/alpha/guide/keras/overview#sequential_model) or with a [custom training loop](https://www.tensorflow.org/alpha/guide/keras/overview#model_subclassing).
+This layer can be used inside a [Keras model](https://www.tensorflow.org/guide/keras/overview#sequential_model) or with a [custom training loop](https://www.tensorflow.org/guide/keras/train_and_evaluate#part_ii_writing_your_own_training_evaluation_loops_from_scratch).
 
 ## Examples
 
@@ -40,7 +40,7 @@ Check out our examples on how to train a Binarized Neural Network in just a few 
 
 Before installing Larq, please install:
 
-- [Python](https://python.org) version `3.6` or `3.7`
+- [Python](https://www.python.org/) version `3.6` or `3.7`
 - [Tensorflow](https://www.tensorflow.org/install) version `1.13`, `1.14`, `1.15` or `2.0.0`:
   ```shell
   pip install tensorflow  # or tensorflow-gpu
