### 2019-10-14 9abccce
Fix links on front page (#288)
This changes the links to the docs of the stable release and resolves the redirect in the link to the Python website.
