### 2019-10-15 ce77fbd
Use animated logo in the README
