### 2019-10-16 8eba852
activations_test update
