+++ larq/activations_test.py
@@ -2,6 +2,7 @@ import tensorflow as tf
 import numpy as np
 import pytest
 import larq as lq
+from larq.testing_utils import generate_real_values_with_zeros
 
 
 @pytest.mark.parametrize("name", ["hard_tanh", "leaky_tanh"])
@@ -15,7 +16,7 @@ def test_serialization(name):
 
 
 def test_hard_tanh():
-    real_values = np.random.uniform(-2, 2, (3, 3, 32))
+    real_values = generate_real_values_with_zeros()
     x = tf.keras.backend.placeholder(ndim=2)
     f = tf.keras.backend.function([x], [lq.activations.hard_tanh(x)])
     result = f([real_values])[0]
@@ -24,16 +25,16 @@ def test_hard_tanh():
 
 def test_leaky_tanh():
     @np.vectorize
-    def leaky_tanh(x):
+    def leaky_tanh(x, alpha):
         if x <= -1:
-            return -1 + 0.2 * (x + 1)
+            return -1 + alpha * (x + 1)
         elif x <= 1:
             return x
         else:
-            return 1 + 0.2 * (x - 1)
+            return 1 + alpha * (x - 1)
 
-    real_values = np.random.uniform(-2, 2, (3, 3, 32))
+    real_values = generate_real_values_with_zeros()
     x = tf.keras.backend.placeholder(ndim=2)
     f = tf.keras.backend.function([x], [lq.activations.leaky_tanh(x)])
     result = f([real_values])[0]
-    np.testing.assert_allclose(result, leaky_tanh(real_values))
+    np.testing.assert_allclose(result, leaky_tanh(real_values, alpha=0.2))
