### 2019-10-16 a9c26fa
Cleanup lq.math tests (#292)
* math_test

* activations_test

* constraints_test
