### 2019-10-17 53fef63
Heaviside Function
