### 2019-10-17 d4b25ea
Test against TensorFlow 1.15.0 stable
