+++ .github/workflows/unittest.yml
@@ -7,7 +7,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        tf-version: [1.13.1, 1.14.0, 1.15.0-rc2, 2.0.0]
+        tf-version: [1.13.1, 1.14.0, 1.15.0, 2.0.0]
 
     steps:
       - uses: actions/checkout@v1
