### 2019-10-17 c677edf
Add google analytics tracking (#295)
I disabled all adds features in Google analytics and [mkdocs material anonymises the IPs by default](https://github.com/squidfunk/mkdocs-material/blob/91f2289587c4340f1f8cb515ef3383ec9888757e/material/partials/integrations/analytics.html#L12) so we should be compliant with GDPR even without specifically asking users for permission.
