### 2019-10-21 828fb19
Bump pillow from 6.2.0 to 6.2.1
Bumps [pillow](https://github.com/python-pillow/Pillow) from 6.2.0 to 6.2.1.
- [Release notes](https://github.com/python-pillow/Pillow/releases)
- [Changelog](https://github.com/python-pillow/Pillow/blob/master/CHANGES.rst)
- [Commits](https://github.com/python-pillow/Pillow/compare/6.2.0...6.2.1)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
