### 2019-10-24 48b2012
Bump tensorflow-datasets from 1.2.0 to 1.3.0 (#299)
Bumps [tensorflow-datasets](https://github.com/tensorflow/datasets) from 1.2.0 to 1.3.0.
- [Release notes](https://github.com/tensorflow/datasets/releases)
- [Changelog](https://github.com/tensorflow/datasets/blob/master/docs/release_notes.md)
- [Commits](https://github.com/tensorflow/datasets/commits)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
