### 2019-10-28 7d727cb
Move all flip_ratio params into tf.init_scope
