### 2019-10-30 28cfdfa
Add quotes around project setup pip commands
This makes them work in both bash and zsh.
