### 2019-10-30 042210d
Also add quotes to documentation pip install
