### 2019-10-30 b474258
Enhancements for CONTRIBUTING.md shell commands + adding editors to .gitignore (#303)
* Ignore .vscode/ folder

* Ignore .idea/ folder

* Add quotes around project setup pip commands

This makes them work in both bash and zsh.

* Also add quotes to documentation pip install
