### 2019-10-30 34538a5
Fix zero division bug in summary for weightless models (#302)
* Fix zero division bug in summary for weightless models

* Add unit test for weightless model summary
