### 2019-11-01 ee39cdf
Refactor Bop into Bop + OptimizerGroup
