### 2019-11-01 75b56b7
WIP: Implement Quantized Variable
