+++ larq/quantized_variable.py
@@ -1,0 +1,423 @@
+"""Contains QuantizedVariable, a variable which automatically casts itself."""
+from tensorflow.python.distribute import values as distribute_values
+from tensorflow.python.eager import context
+from tensorflow.python.framework import ops
+from tensorflow.python.ops import resource_variable_ops
+from tensorflow.python.ops import variables
+from larq.quantized_scope import should_quantize
+
+
+class QuantizedVariable(variables.Variable):
+    """Variable that will cast itself to a different dtype in applicable contexts.
+
+  This class wraps a floating-point tf.Variable. It emulates the variable
+  interface and delegates to the wrapped variable, but it additionally will cast
+  the wrapped variable under a `Graph._enable_variable_auto_cast(dtype)` context
+  manager.
+
+  For example:
+
+  ```
+  v = tf.Variable(1.0, dtype=tf.float32)
+  v = QuantizedVariable(v)
+  print(tf.identity(v).dtype)  # tf.float32
+  with ops.get_default_graph()._enable_variable_auto_cast(tf.float16):
+    print(tf.identity(v).dtype)  # tf.float16, as v will cast itself to float16
+    print(v.dtype)  # tf.float16, as v.dtype also changes under the ctx manager.
+  ```
+
+  The purpose of this class is to allow Keras layers to create variables in
+  float32, and automatically cast them to float16 or bfloat16 when the layer is
+  called.
+  """
+
+    def __init__(self, variable, quantizer=None):
+        """Creates an QuantizedVariable instance.
+
+        Args:
+            variable: A floating-point resource variable to wrap.
+
+        Raises:
+            ValueError: If `variable` is not a floating-point resource variable
+        """
+        if not resource_variable_ops.is_resource_variable(variable):
+            raise ValueError(
+                "variable must be of type tf.ResourceVariable, but got: "
+                "%s" % variable
+            )
+        self._variable = variable
+        self.quantizer = quantizer
+
+    @property
+    def precision(self):
+        try:
+            return self.quantizer.precision
+        except:
+            return 32
+
+    def _should_quantize(self):
+        """Returns True if this variable should be quantized when accessed."""
+        if self.quantizer is not None:
+            return should_quantize()
+        return False
+
+    @property
+    def dtype(self):
+        """The dtype this variable will be casted to when read."""
+        return self._variable.dtype
+
+    def _as_graph_element(self):
+        """Conversion function for Graph.as_graph_element()."""
+        return self._variable
+
+    def value(self):
+        val = self._variable.value()
+        if not self._should_quantize():
+            return val
+        return self.quantizer(val)
+
+    def read_value(self):
+        val = self._variable.read_value()
+        # TODO: Should this check be removed? see upstream
+        if not self._should_quantize():
+            return val
+        return self.quantizer(val)
+
+    def sparse_read(self, indices, name=None):
+        """Reads the value of this variable sparsely, using `gather`."""
+        val = self._variable.sparse_read(indices, name=name)
+        # TODO: Should this check be removed? see upstream
+        if not self._should_quantize():
+            return val
+        return self.quantizer(val)
+
+    def gather_nd(self, indices, name=None):
+        """Gather slices of the variable into a Tensor."""
+        val = self._variable.gather_nd(indices, name=name)
+        # TODO: Should this check be removed? see upstream
+        if not self._should_quantize():
+            return val
+        return self.quantizer(val)
+
+    def __getattr__(self, name):
+        return getattr(self._variable, name)
+
+    def _dense_var_to_tensor(self, dtype=None, name=None, as_ref=False):
+        """Converts this variable to a tensor."""
+        if not self._should_quantize():
[…]
