### 2019-11-02 1501be4
Use internal_convert_to_tensor to support as_ref in earlier TF versions
