+++ larq/quantized_variable.py
@@ -105,7 +105,7 @@ class QuantizedVariable(variables.Variable):
     def _dense_var_to_tensor(self, dtype=None, name=None, as_ref=False):
         """Converts this variable to a tensor."""
         if not self._should_quantize():
-            return ops.convert_to_tensor(self._variable, dtype, name, as_ref)
+            return ops.internal_convert_to_tensor(self._variable, dtype, name, as_ref)
         # TODO(reedwm): Support as_ref?
         assert not as_ref
         if dtype is not None and not dtype.is_compatible_with(self.dtype):
@@ -113,7 +113,7 @@ class QuantizedVariable(variables.Variable):
                 "Incompatible type conversion requested to type {!r} for variable "
                 "of type {!r}".format(dtype.name, self.dtype.name)
             )
-        val = ops.convert_to_tensor(
+        val = ops.internal_convert_to_tensor(
             self._variable, self._variable.dtype, name, as_ref=False
         )
         return self.quantizer(val)
