### 2019-11-04 4ce2e10
Fix refactor syntax issues
