### 2019-11-04 4590e36
Update Bop invocations in optimizer tests
