### 2019-11-04 29f3e8c
Fix name getter in Bop serializer
