### 2019-11-04 f65d7a1
Rename OptimizerGroup to BNNOptimizerDuo
