+++ larq/optimizers.py
@@ -1,7 +1,7 @@
 from larq import utils
 
 if utils.tf_1_14_or_newer():
-    from larq.optimizers_v2 import Bop, OptimizerGroup
+    from larq.optimizers_v2 import Bop, BNNOptimizerDuo
 
     __all__ = ["Bop", "OptimizerGroup"]
 else:
