### 2019-11-04 3dd59a8
Remove _get_decayed_hyper from BNNOptimizerDuo
