### 2019-11-04 51a20ab
Remove _is_binary check from _create_slots in Bop
