### 2019-11-04 bed1b50
Reorder methods in BNNOptimizerDuo
