### 2019-11-04 8968fce
Implement serialization and deserialization
