### 2019-11-04 1cbf94f
Fix __all__ in optimizers.py
