### 2019-11-04 001fdbd
Move __all__ from optimizers.py into _v1.py and _v2.py
