### 2019-11-04 aa8ebad
Revert "Move __all__ from optimizers.py into _v1.py and _v2.py"
This reverts commit 001fdbdf5630f415072f74e4b97cfaccea635aa9.
