### 2019-11-04 27b4fed
Fix and simplify string formatting
