### 2019-11-04 628d227
Remove mentions of AutocastVariable and pylint comments
