### 2019-11-05 6f8732d
Remove leftover _as_graph_element method
