### 2019-11-05 1ee0064
Refactor using MaskedOptimizer and OptimizerGroup
