### 2019-11-05 be28872
Small docstring fixes
