### 2019-11-05 bb07540
Clean up apply_gradients
