### 2019-11-05 5de03e3
Update tests to use new API
