### 2019-11-05 62c28b0
Refactor using tuples and CaseOptimizer
