### 2019-11-06 f1eb04f
Add callable() check for predicates
