### 2019-11-06 151af08
Throw error for getting lr from CaseOptimizer
