### 2019-11-06 678590c
Throw warning instead of error on missing default optimizer
