### 2019-11-06 1b49c96
Fix __all__ in optimizers.py
