+++ larq/optimizers.py
@@ -3,7 +3,7 @@ from larq import utils
 if utils.tf_1_14_or_newer():
     from larq.optimizers_v2 import Bop, CaseOptimizer
 
-    __all__ = ["Bop", "OptimizerGroup"]
+    __all__ = ["Bop", "CaseOptimizer"]
 else:
     from larq.optimizers_v1 import Bop, XavierLearningRateScaling
 
