### 2019-11-06 6e6ccd8
Improve documentation
