### 2019-11-06 5eb2710
Add default and tweak name in CaseOptimizer init
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
