### 2019-11-06 b33bd04
Improve legibility of train_ops generation
Co-Authored-By: Daniël Vos <daniel-vos@users.noreply.github.com>
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
