+++ larq/optimizers_v2.py
@@ -77,7 +77,7 @@ class CaseOptimizer(tf.keras.optimizers.Optimizer):
         optimizers = [opt for (_, opt) in self.pred_opt_pairs]
         train_ops = [
             optimizer.apply_gradients(grads_and_vars, name=name)
-            for operimizer, grads_and_vars in zip(optimizers, opt_grads_and_vars)
+            for optimizer, grads_and_vars in zip(optimizers, opt_grads_and_vars)
         ]
 
         if len(default_grads_and_vars) > 0 and self.default is None:
