### 2019-11-06 bb67d9f
Implement pre-computing variable to gradient mapping
