### 2019-11-06 ef4a646
Remove old code and clean up comments
