### 2019-11-06 7aa8624
Replace mapper yielding approach with list
