### 2019-11-06 253c5d7
Rename is_binary to is_binary_variable
