### 2019-11-06 38b315d
Remove __getattr__ from CaseOptimizer
