### 2019-11-06 87bcf8e
Update optimizer API in tests
