### 2019-11-06 992350a
Implement serialization and deserialization
