### 2019-11-06 ba2ca89
Tweak docs and variable names
