### 2019-11-06 c5a41fc
:arrow_up: MathJax@3
This reverts #228 and fixes inconsistent MathJax syntax since MathJax 3 uses stricter defaults.
This should speedup our math rendering.
