### 2019-11-07 b077be9
Run unit tests in parallel using `pytest-xdist`
This should provide better isolation between different tests and might speed up CI slightly.
