### 2019-11-07 1f477e0
Add documentation for larq.optimizers
