### 2019-11-07 9f477c8
Small docs tweaks
