+++ larq/optimizers.py
@@ -1,14 +1,14 @@
 """Neural networks with extremely low-precision weights and activations, such as 
-Binarized Neural Networks (BNNs), usually contain a mix of low-precision and 
-higher-precision weights (e.g. 8-bit, 16-bit, or 32-bit). Examples of this include the 
-first and last layers of image classificiation models, which have higher-precision 
-weights in most BNN architectures from the literature.
+Binarized Neural Networks (BNNs), usually contain a mix of low-precision weights (e.g. 
+1-bit) and  higher-precision weights (e.g. 8-bit, 16-bit, or 32-bit). Examples of this 
+include the first and last layers of image classificiation models, which have 
+higher-precision weights in most BNN architectures from the literature.
 
 Optimizing a BNN, then, consists of optimizing both low-precision and higher-precision
 weights. In `larq`, we provide a mechanism to target different bit-precision variables
 with different optimizers using the `CaseOptimizer` class. Modeled after the
 [`tf.case`](https://www.tensorflow.org/api_docs/python/tf/case) signature,
-`CaseOptimizer` accepts pairs of optimizers and predicates. A predicate, given a
+`CaseOptimizer` accepts pairs of predicates and optimizers. A predicate, given a
 variable, decides whether its optimizer should train that variable.
 
 A `CaseOptimizer` behaves much like any other
