### 2019-11-07 d9665ed
Explain grads_and_vars cast to list
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
