### 2019-11-07 8ae8906
Fix pred_opt_pairs argument documentation
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
