### 2019-11-07 be7c8d7
Small docs tweak
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
