### 2019-11-07 ab17a69
Remove name from Bop serialization
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
