### 2019-11-07 ef4dd04
Small docs tweak
Co-Authored-By: Lukas Geiger <lgeiger@users.noreply.github.com>
