### 2019-11-07 2ec9175
Allow passing one or several tuples instead of list
