+++ larq/optimizers_test.py
@@ -114,16 +114,16 @@ class TestBopOptimizer:
     def test_bop_accuracy(self):
         _test_optimizer(
             lq.optimizers.CaseOptimizer(
-                [(lq.optimizers.Bop.is_binary_variable, lq.optimizers.Bop())],
-                tf.keras.optimizers.Adam(0.01),
+                (lq.optimizers.Bop.is_binary_variable, lq.optimizers.Bop()),
+                default=tf.keras.optimizers.Adam(0.01),
             ),
             test_kernels_are_binary=True,
         )
         # test optimizer on model with only binary trainable vars (low accuracy)
         _test_optimizer(
             lq.optimizers.CaseOptimizer(
-                [(lq.optimizers.Bop.is_binary_variable, lq.optimizers.Bop())],
-                tf.keras.optimizers.Adam(0.01),
+                (lq.optimizers.Bop.is_binary_variable, lq.optimizers.Bop()),
+                default=tf.keras.optimizers.Adam(0.01),
             ),
             test_kernels_are_binary=True,
             trainable_bn=False,
@@ -137,20 +137,18 @@ class TestBopOptimizer:
     def test_bop_tf_1_14_schedules(self):
         _test_optimizer(
             lq.optimizers.CaseOptimizer(
-                [
-                    (
-                        lq.optimizers.Bop.is_binary_variable,
-                        lq.optimizers.Bop(
-                            threshold=tf.keras.optimizers.schedules.InverseTimeDecay(
-                                3.0, decay_steps=1.0, decay_rate=0.5
-                            ),
-                            gamma=tf.keras.optimizers.schedules.InverseTimeDecay(
-                                3.0, decay_steps=1.0, decay_rate=0.5
-                            ),
+                (
+                    lq.optimizers.Bop.is_binary_variable,
+                    lq.optimizers.Bop(
+                        threshold=tf.keras.optimizers.schedules.InverseTimeDecay(
+                            3.0, decay_steps=1.0, decay_rate=0.5
                         ),
-                    )
-                ],
-                tf.keras.optimizers.Adam(0.01),
+                        gamma=tf.keras.optimizers.schedules.InverseTimeDecay(
+                            3.0, decay_steps=1.0, decay_rate=0.5
+                        ),
+                    ),
+                ),
+                default=tf.keras.optimizers.Adam(0.01),
             ),
             test_kernels_are_binary=True,
         )
@@ -167,8 +165,8 @@ class TestBopOptimizer:
         model.compile(
             loss="categorical_crossentropy",
             optimizer=lq.optimizers.CaseOptimizer(
-                [(lq.optimizers.Bop.is_binary_variable, lq.optimizers.Bop())],
-                tf.keras.optimizers.Adam(0.01),
+                (lq.optimizers.Bop.is_binary_variable, lq.optimizers.Bop()),
+                default=tf.keras.optimizers.Adam(0.01),
             ),
         )
 
@@ -187,7 +185,7 @@ class TestBopOptimizer:
     def test_bop_serialization(self):
         _test_serialization(
             lq.optimizers.CaseOptimizer(
-                [(lq.optimizers.Bop.is_binary_variable, lq.optimizers.Bop())],
-                tf.keras.optimizers.Adam(0.01),
+                (lq.optimizers.Bop.is_binary_variable, lq.optimizers.Bop()),
+                default=tf.keras.optimizers.Adam(0.01),
             ),
         )
