### 2019-11-07 28d6813
Remove tests relying on optimizer callbacks
We no longer support these with the new CaseOptimizer setup
