### 2019-11-07 0ff829a
Remove unused imports from callback_test
