### 2019-11-07 b28352b
Fix default optimizer type check
