### 2019-11-07 f27aeed
Assert imports in optimizers to make pyflakes happy
