### 2019-11-07 9cd01e3
Call each optimizer only once per apply_gradients
