### 2019-11-07 35b3d36
Small docs tweaks
Co-Authored-By: Koen Helwegen <koenhelwegen@users.noreply.github.com>
