### 2019-11-07 fafa64e
Use tf.test.TestCase instead of keras_parameterized
This was added to fix some issues we had with different tests interfering with each other. This is now fixed since we use `pytest-xdist` to provide better isolation.
