### 2019-11-07 928a875
Skip CaseOptimizer tests for TF <= 1.13
