### 2019-11-08 6331bb6
Add pytest fixture to run tests in eager and graph mode
