+++ larq/metrics_test.py
@@ -1,6 +1,5 @@
 import tensorflow as tf
 from larq import metrics
-from tensorflow.python.keras import keras_parameterized
 import pytest
 
 
@@ -14,70 +13,69 @@ def test_scope():
             pass
 
 
-class FlipRatioTest(keras_parameterized.TestCase):
-    def test_config(self):
-        mcv = metrics.FlipRatio(
-            values_shape=[3, 3], values_dtype="int16", name="mcv", dtype=tf.float16
+def test_config():
+    mcv = metrics.FlipRatio(
+        values_shape=[3, 3], values_dtype="int16", name="mcv", dtype=tf.float16
+    )
+    assert mcv.name == "mcv"
+    assert mcv.stateful
+    assert mcv.dtype == tf.float16
+    assert mcv.values_dtype == tf.int16
+    assert mcv.values_shape == [3, 3]
+
+    mcv2 = metrics.FlipRatio.from_config(mcv.get_config())
+    assert mcv2.name == "mcv"
+    assert mcv2.stateful
+    assert mcv2.dtype == tf.float16
+    assert mcv2.values_dtype == tf.int16
+    assert mcv2.values_shape == [3, 3]
+
+
+def test_metric(eager_mode):
+    mcv = metrics.FlipRatio([2])
+
+    assert 0 == mcv.result().numpy()
+    assert 0 == mcv.total.numpy()
+    assert 0 == mcv.count.numpy()
+
+    mcv.update_state([1, 1])
+    assert all([1, 1] == mcv._previous_values.numpy())
+    assert 0 == mcv.total.numpy()
+    assert 1 == mcv.count.numpy()
+    assert 0 == mcv.result().numpy()
+
+    mcv.update_state([2, 2])
+    assert all([2, 2] == mcv._previous_values.numpy())
+    assert 1 == mcv.total.numpy()
+    assert 2 == mcv.count.numpy()
+    assert 1 == mcv.result().numpy()
+
+    mcv.update_state([1, 2])
+    assert all([1, 2] == mcv._previous_values.numpy())
+    assert 1.5 == mcv.total.numpy()
+    assert 3 == mcv.count.numpy()
+    assert 1.5 / 2 == mcv.result().numpy()
+
+
+def test_metric_in_graph_mode(graph_mode):
+    mcv = metrics.FlipRatio([2])
+
+    new_state = tf.compat.v1.placeholder(dtype=tf.float32, shape=[2])
+    update_state_op = mcv.update_state(new_state)
+    metric_value = mcv.result()
+
+    with tf.compat.v1.Session() as sess:
+        sess.run(tf.compat.v1.variables_initializer(mcv.variables))
+
+        sess.run(update_state_op, feed_dict={new_state: [1, 1]})
+        sess.run(update_state_op, feed_dict={new_state: [2, 2]})
+        sess.run(update_state_op, feed_dict={new_state: [1, 2]})
+
+        previous, total, count, result = sess.run(
+            [mcv._previous_values, mcv.total, mcv.count, metric_value]
         )
-        self.assertEqual(mcv.name, "mcv")
-        self.assertTrue(mcv.stateful)
-        self.assertEqual(mcv.dtype, tf.float16)
-        self.assertEqual(mcv.values_dtype, tf.int16)
-        self.assertEqual(mcv.values_shape, [3, 3])
 
-        mcv2 = metrics.FlipRatio.from_config(mcv.get_config())
-        self.assertEqual(mcv2.name, "mcv")
-        self.assertTrue(mcv2.stateful)
-        self.assertEqual(mcv2.dtype, tf.float16)
-        self.assertEqual(mcv2.values_dtype, tf.int16)
[…]
