### 2019-11-08 ce81f20
Support depthwise and separable layers
This is hopefully a first step to reduce duplication in the base layers
