### 2019-11-08 c1d472c
Remove quantized_latent_weights and quantizer arguments
since they are not needed anymore with the new API
