### 2019-11-08 ef6610c
Move non_trainable_weights to base class
