### 2019-11-08 93cd1ff
Add pytest fixture to run tests in eager and graph mode (#313)
