### 2019-11-11 97ef8f8
Clean up optimizer tests
