### 2019-11-11 39ad33f
Add CaseOptimizer type check tests
