### 2019-11-11 eeb5272
Add variable optimizer mapping tests
