### 2019-11-11 e4695d2
Make CaseOptimizer signature more descriptive
