### 2019-11-11 cf6491f
Drop support for TensorFlow 1.13
