+++ larq/optimizers.py
@@ -30,18 +30,254 @@ optimizer. A variable may not be claimed by more than one optimizer's predicate.
     ```
 """
 
+
+import warnings
+
+import tensorflow as tf
+import larq as lq
+
 from larq import utils
+from copy import deepcopy
 
-if utils.tf_1_14_or_newer():
-    from larq.optimizers_v2 import Bop, CaseOptimizer
 
-    __all__ = ["Bop", "CaseOptimizer"]
-    assert Bop  # Asserts keep pyflakes happy
-    assert CaseOptimizer
+__all__ = ["Bop", "CaseOptimizer"]
 
-else:
-    from larq.optimizers_v1 import Bop, XavierLearningRateScaling
 
-    __all__ = ["Bop", "XavierLearningRateScaling"]
-    assert Bop
-    assert XavierLearningRateScaling
+@utils.register_keras_custom_object
+class CaseOptimizer(tf.keras.optimizers.Optimizer):
+    """An optmizer wrapper that applies different optimizers to a subset of variables.
+
+    An optimizer is used to train a variable iff its accompanying predicate evaluates to
+    `True`.
+
+    For each variable, at most one optimizer's predicate may evaluate to `True`. If no
+    optimizer's predicate evaluates to `True` for a variable, it is trained with the
+    `default_optimizer`. If a variable is claimed by no optimizers and
+    `default_optimizer == None`, the variable is not trained.
+
+    # Arguments
+    predicate_optimizer_pairs: One or more `(pred, tf.keras.optimizers.Optimzer)` pairs,
+        where `pred`  takes one `tf.Variable` as argument and returns `True` if the
+        optimizer should be used for that variable, e.g. `pred(var) == True`.
+    default_optimizer: A `tf.keras.optimizers.Optimizer` to be applied to any variable
+        not claimed by any other optimizer. (Must be passed as keyword argument.)
+    """
+
+    def __init__(
+        self, *predicate_optimizer_pairs, default_optimizer=None, name="optimizer_case"
+    ):
+        super().__init__(name=name)
+
+        # Type checks for (predicate, optimizer) pairs
+        for i, (predicate, optimizer) in enumerate(predicate_optimizer_pairs):
+            if not callable(predicate):
+                raise TypeError(
+                    f"Expected callable predicate at `predicate_optimizer_pairs[{i}][0]` but got `{type(predicate)}`."
+                )
+            if not isinstance(optimizer, tf.keras.optimizers.Optimizer):
+                raise TypeError(
+                    f"Expected `tf.keras.optimizers.Optimizer` at `predicate_optimizer_pairs[{i}][1]` but got `{type(optimizer)}`."
+                )
+
+        # Type check for default optimizers
+        if default_optimizer is not None and not isinstance(
+            default_optimizer, tf.keras.optimizers.Optimizer
+        ):
+            raise TypeError(
+                f"Expected `tf.keras.optimizers.Optimizer` for `default_optimizer` but got `{type(default_optimizer)}`."
+            )
+
+        self.pred_opt_pairs = predicate_optimizer_pairs
+        self.default = default_optimizer
+
+        self.var_opt_mapping = None
+
+        # List of optimizers ending in `default_optimizer`, for easier internal access
+        self.optimizers = [opt for (_, opt) in self.pred_opt_pairs]
+
+        if self.default:
+            self.optimizers.append(self.default)
+            self.DEFAULT_OPT_INDEX = len(self.pred_opt_pairs)
+
+    def apply_gradients(self, grads_and_vars, name=None):
+        """Apply gradients to variables for each optimizer.
+
+        On the first call to `apply_gradients()`, compute the mapping from variables to
+        optimizers and cache it in the `self.var_opt_mapping` dict for serialization and
+        faster access.
+        """
+
+        if self.var_opt_mapping is None:
[…]
