+++ README.md
@@ -41,7 +41,7 @@ Check out our examples on how to train a Binarized Neural Network in just a few 
 Before installing Larq, please install:
 
 - [Python](https://www.python.org/) version `3.6` or `3.7`
-- [Tensorflow](https://www.tensorflow.org/install) version `1.13`, `1.14`, `1.15` or `2.0.0`:
+- [Tensorflow](https://www.tensorflow.org/install) version `1.14`, `1.15` or `2.0.0`:
   ```shell
   pip install tensorflow  # or tensorflow-gpu
   ```

[…]
