### 2019-11-11 32b6c2e
Drop support for TensorFlow 1.13
