### 2019-11-11 362652e
Add formatting tests
Easy lines to make codecov happy
