### 2019-11-11 8351e30
Refactor Bop into Bop + CaseOptimizer (#304)
* Refactor Bop into Bop + OptimizerGroup

* Fix refactor syntax issues

* Update Bop invocations in optimizer tests

* Fix name getter in Bop serializer

* Rename OptimizerGroup to BNNOptimizerDuo

* Remove _get_decayed_hyper from BNNOptimizerDuo

* Remove _is_binary check from _create_slots in Bop

* Reorder methods in BNNOptimizerDuo

* Implement serialization and deserialization

* Remove old TODOs

* Fix __all__ in optimizers.py

* Move __all__ from optimizers.py into _v1.py and _v2.py

* Revert "Move __all__ from optimizers.py into _v1.py and _
[…]
