+++ apidocs.yml
@@ -8,6 +8,7 @@
     - larq.constraints+
 - optimizers.md:
     - larq.optimizers:
+        - larq.optimizers_v2.CaseOptimizer
         - larq.optimizers_v2.Bop
         - larq.optimizers_v1.XavierLearningRateScaling
 - math.md:

[…]
