### 2019-11-11 afdc210
Fix log count check in layer warnings tests (#316)
