+++ larq/layers_test.py
@@ -189,7 +189,7 @@ class LayersTest(keras_parameterized.TestCase):
 
 def test_layer_warns(caplog):
     lq.layers.QuantDense(5, kernel_quantizer="ste_sign")
-    assert len(caplog.records) == 1
+    assert len(caplog.records) >= 1
     assert "kernel_constraint" in caplog.text
 
 
@@ -202,7 +202,7 @@ def test_layer_does_not_warn(caplog):
 
 def test_depthwise_layer_warns(caplog):
     lq.layers.QuantDepthwiseConv2D(5, depthwise_quantizer="ste_sign")
-    assert len(caplog.records) == 1
+    assert len(caplog.records) >= 1
     assert "depthwise_constraint" in caplog.text
 
 
