### 2019-11-11 6c9c691
Remove LarqMetric
