### 2019-11-11 310ace2
Simplify quantizers.py imports
