### 2019-11-11 ee29ae3
Drop TensorFlow 1.13  (#317)
* Drop support for TensorFlow 1.13

* Fix pyflakes

* Fix docs

* Add formatting tests

Easy lines to make codecov happy

* Remove LarqMetric

* Simplify quantizers.py imports
