### 2019-11-11 dd1e220
Add initial version of quantized variable
