+++ larq/quantized_variable.py
@@ -1,0 +1,290 @@
+"""Contains QuantizedVariable, a variable that can be quantized in the forward pass."""
+from tensorflow.python.distribute import values as distribute_values
+from tensorflow.python.eager import context
+from tensorflow.python.framework import ops
+from tensorflow.python.ops import resource_variable_ops
+from tensorflow.python.ops import variables
+from functools import wraps
+
+
+def quantize(method):
+    """A decorator that can quantize the return value of the classmethod.
+
+    Syntactic sugar for `self.quantizer(self.method(*args, **kwargs))`.
+    """
+
+    @wraps(method)
+    def wrapper(self, *args, **kwargs):
+        value = method(self, *args, **kwargs)
+        if self.quantizer:
+            return self.quantizer(value)
+        return value
+
+    return wrapper
+
+
+class QuantizedVariable(variables.Variable):
+    """Variable that can be quantized in the forward pass in applicable contexts."""
+
+    def __init__(self, variable, quantizer=None, precision=None):
+        """Creates an QuantizedVariable instance.
+
+        # Arguments
+        variable: A floating-point resource variable to wrap.
+        quantizer: An optional quantizer to transform the floating-point
+            variable to a fake quantized variable.
+        precision: An optional integer defining the precision of the quantized
+            variable. If `None`, `quantizer.precision` is used.
+        """
+        if not resource_variable_ops.is_resource_variable(variable):
+            raise ValueError(
+                "variable must be of type tf.ResourceVariable, but got: "
+                "%s" % variable
+            )
+        self.latent_variable = variable
+        self.quantizer = quantizer
+        self.precision = precision or getattr(quantizer, "precision", None)
+
+    @quantize
+    def value(self):
+        return self.latent_variable.value()
+
+    @quantize
+    def read_value(self):
+        return self.latent_variable.read_value()
+
+    @quantize
+    def sparse_read(self, indices, name=None):
+        """Reads the value of this variable sparsely, using `gather`."""
+        return self.latent_variable.sparse_read(indices, name=name)
+
+    @quantize
+    def gather_nd(self, indices, name=None):
+        """Gather slices of the variable into a Tensor."""
+        return self.latent_variable.gather_nd(indices, name=name)
+
+    def __getattr__(self, name):
+        return getattr(self.latent_variable, name)
+
+    @quantize
+    def _dense_var_to_tensor(self, dtype=None, name=None, as_ref=False):
+        """Converts this variable to a tensor."""
+        return self.latent_variable._dense_var_to_tensor(dtype, name, as_ref)
+
+    @quantize
+    def eval(self, session=None):
+        return self.latent_variable.eval(session)
+
+    @quantize
+    def initialized_value(self):
+        return self.latent_variable.initialized_value()
+
+    @property
+    @quantize
+    def initial_value(self):
+        return self.latent_variable.initial_value
+
+    def _should_act_as_resource_variable(self):
+        """Pass resource_variable_ops.is_resource_variable check."""
+        pass
+
+    def __repr__(self):
+        if context.executing_eagerly() and not self._in_graph_mode:
+            return (
+                f"<QuantizedVariable '{self.name}' shape={self.shape} "
+                f"dtype={self.dtype.name} quantizer={self.quantizer.__repr__()} "
+                f"precision={self.precision} "
+                f"numpy={ops.numpy_text(self.read_value(), is_repr=True)}>"
+            )
+        return (
+            f"<QuantizedVariable '{self.name}' shape={self.shape} "
+            f"dtype={self.dtype.name} quantizer={self.quantizer.__repr__()} "
+            f"precision={self.precision}>"
+        )
+
+    # Method delegations: We delegate the following methods to self.latent_variable.
[…]
