### 2019-11-12 e9222a6
Add link to NeurIPS poster (#318)
