### 2019-11-12 6c25dc4
Add distribution scope pytest fixture
