+++ larq/conftest_test.py
@@ -2,10 +2,7 @@ import tensorflow as tf
 
 
 def test_eager_and_graph_mode_fixture(eager_and_graph_mode):
-    if eager_and_graph_mode == "eager":
-        assert tf.executing_eagerly()
-    else:
-        assert not tf.executing_eagerly()
+    assert tf.executing_eagerly() == (eager_and_graph_mode == "eager")
 
 
 def test_eager_mode_fixture(eager_mode):
@@ -14,3 +11,7 @@ def test_eager_mode_fixture(eager_mode):
 
 def test_graph_mode_fixture(graph_mode):
     assert not tf.executing_eagerly()
+
+
+def test_distribute_scope(distribute_scope):
+    assert tf.distribute.has_strategy() == (distribute_scope == "distribute")
