### 2019-11-12 4c216f4
Fix eval
We can't use the quantize decorator here, since eval would return a Tensor instead of an numpy array.
