### 2019-11-12 2b5f15d
Update larq/layers_base.py
Co-Authored-By: Leon Overweel <l.overweel@gmail.com>
