### 2019-11-12 8549ea1
Start refactor with fixture_run_all_keras_modes
