+++ larq/conftest.py
@@ -21,3 +21,16 @@ def eager_and_graph_mode(request):
     """pytest fixture for running test in eager and graph mode"""
     with getattr(context, f"{request.param}_mode")():
         yield request.param
+
+
+@pytest.fixture(params=["tf_eager", "tf_keras_eager", "graph"])
+def fixture_run_all_keras_modes(request):
+    """TODO"""
+
+    if request.param == "graph":
+        with context.graph_mode():
+            yield
+
+    else:
+        with context.eager_mode():
+            yield request.param == "tf_keras_eager"
