### 2019-11-12 6d950dd
Rename and add documentation for keras_should_run_eagerly
