### 2019-11-12 0ac6979
Move test_depthwise_layers into testing class
