### 2019-11-12 4dfca23
Refactor test_separable_layers with keras_should_run_eagerly
