### 2019-11-12 784d743
Refactor test_binarization with keras_should_run_eagerly
