+++ larq/conftest.py
@@ -1,4 +1,5 @@
 import pytest
+import tensorflow as tf
 from tensorflow.python.eager import context
 
 
@@ -37,6 +38,9 @@ def keras_should_run_eagerly(request):
     """
 
     if request.param == "graph":
+        if int(tf.__version__[0]) >= 2:
+            pytest.skip("Skipping graph mode for TensorFlow 2+.")
+
         with context.graph_mode():
             yield
 
