### 2019-11-12 1e4fc38
Move layer warnings tests into TestLayerWarns class
