### 2019-11-12 551f557
Remove direct dependency on absl-py
