### 2019-11-12 cee09d0
Refactor tests to not use TensorFlow TestCase (#319)
* Start refactor with fixture_run_all_keras_modes

* Rename and add documentation for keras_should_run_eagerly

* Move test_depthwise_layers into testing class

* Refactor test_separable_layers with keras_should_run_eagerly

* Refactor test_binarization with keras_should_run_eagerly

* Move layer warnings tests into TestLayerWarns class

* Remove now-unused imports 🎉

* Remove direct dependency on absl-py
