+++ larq/layers_test.py
@@ -1,11 +1,61 @@
 import tensorflow as tf
 import numpy as np
-from absl.testing import parameterized
 import larq as lq
 import pytest
 import inspect
 from larq import testing_utils
-from tensorflow.python.keras import keras_parameterized
+
+
+PARAMS_ALL_LAYERS = [
+    (lq.layers.QuantDense, tf.keras.layers.Dense, (3, 2), dict(units=3)),
+    (
+        lq.layers.QuantConv1D,
+        tf.keras.layers.Conv1D,
+        (2, 3, 7),
+        dict(filters=2, kernel_size=3),
+    ),
+    (
+        lq.layers.QuantConv2D,
+        tf.keras.layers.Conv2D,
+        (2, 3, 7, 6),
+        dict(filters=2, kernel_size=3),
+    ),
+    (
+        lq.layers.QuantConv3D,
+        tf.keras.layers.Conv3D,
+        (2, 3, 7, 6, 5),
+        dict(filters=2, kernel_size=3),
+    ),
+    (
+        lq.layers.QuantConv2DTranspose,
+        tf.keras.layers.Conv2DTranspose,
+        (2, 3, 7, 6),
+        dict(filters=2, kernel_size=3),
+    ),
+    (
+        lq.layers.QuantConv3DTranspose,
+        tf.keras.layers.Conv3DTranspose,
+        (2, 3, 7, 6, 5),
+        dict(filters=2, kernel_size=3),
+    ),
+    (
+        lq.layers.QuantLocallyConnected1D,
+        tf.keras.layers.LocallyConnected1D,
+        (2, 8, 5),
+        dict(filters=4, kernel_size=3),
+    ),
+    (
+        lq.layers.QuantLocallyConnected2D,
+        tf.keras.layers.LocallyConnected2D,
+        (8, 6, 10, 4),
+        dict(filters=3, kernel_size=3),
+    ),
+]
+
+PARAMS_SEP_LAYERS = [
+    (lq.layers.QuantSeparableConv1D, tf.keras.layers.SeparableConv1D, (2, 3, 7),),
+    (lq.layers.QuantSeparableConv2D, tf.keras.layers.SeparableConv2D, (2, 3, 7, 6),),
+]
 
 
 def random_input(shape):
@@ -16,64 +66,13 @@ def random_input(shape):
     return data.astype("float32")
 
 
-parameterized_all_layers = parameterized.named_parameters(
-    ("QuantDense", lq.layers.QuantDense, tf.keras.layers.Dense, (3, 2), dict(units=3)),
-    (
-        "QuantConv1D",
-        lq.layers.QuantConv1D,
-        tf.keras.layers.Conv1D,
-        (2, 3, 7),
-        dict(filters=2, kernel_size=3),
-    ),
-    (
-        "QuantConv2D",
-        lq.layers.QuantConv2D,
-        tf.keras.layers.Conv2D,
-        (2, 3, 7, 6),
-        dict(filters=2, kernel_size=3),
-    ),
-    (
-        "QuantConv3D",
-        lq.layers.QuantConv3D,
-        tf.keras.layers.Conv3D,
-        (2, 3, 7, 6, 5),
-        dict(filters=2, kernel_size=3),
-    ),
-    (
-        "QuantConv2DTranspose",
-        lq.layers.QuantConv2DTranspose,
-        tf.keras.layers.Conv2DTranspose,
-        (2, 3, 7, 6),
-        dict(filters=2, kernel_size=3),
-    ),
-    (
-        "QuantConv3DTranspose",
-        lq.layers.QuantConv3DTranspose,
-        tf.keras.layers.Conv3DTranspose,
-        (2, 3, 7, 6, 5),
[…]
