+++ setup.py
@@ -27,7 +27,6 @@ setup(
         "tensorflow": ["tensorflow>=1.14.0"],
         "tensorflow_gpu": ["tensorflow-gpu>=1.14.0"],
         "test": [
-            "absl-py==0.8.1",
             "pytest==5.2.2",
             "pytest-cov==2.8.1",
             "pytest-xdist==1.30.0",
