### 2019-11-13 d1dc165
Simplify layer tests
This removes unnecessary computation from the layer tests which should result in a speed up.
