+++ larq/layers_test.py
@@ -1,4 +1,5 @@
 import tensorflow as tf
+from tensorflow.keras.initializers import constant as const_init
 import numpy as np
 import larq as lq
 import pytest
@@ -83,25 +84,23 @@ class TestLayers:
                     **kwargs,
                     kernel_quantizer="ste_sign",
                     input_quantizer="ste_sign",
-                    kernel_initializer=tf.keras.initializers.constant(random_weight),
+                    kernel_initializer=const_init(random_weight),
                 ),
                 input_data=input_data,
                 should_run_eagerly=keras_should_run_eagerly,
             )
 
-        fp_output = testing_utils.layer_test(
-            layer,
-            kwargs=dict(
-                **kwargs,
-                kernel_initializer=tf.keras.initializers.constant(
-                    np.sign(random_weight)
-                ),
-            ),
-            input_data=np.sign(input_data),
-            should_run_eagerly=keras_should_run_eagerly,
+        fp_model = tf.keras.models.Sequential(
+            [
+                layer(
+                    **kwargs,
+                    kernel_initializer=const_init(np.sign(random_weight)),
+                    input_shape=input_shape[1:],
+                )
+            ]
         )
 
-        np.testing.assert_allclose(quant_output, fp_output)
+        np.testing.assert_allclose(quant_output, fp_model.predict(np.sign(input_data)))
 
     @pytest.mark.parametrize("quantized_layer, layer, input_shape", PARAMS_SEP_LAYERS)
     def test_separable_layers(
@@ -120,34 +119,26 @@ class TestLayers:
                     depthwise_quantizer="ste_sign",
                     pointwise_quantizer="ste_sign",
                     input_quantizer="ste_sign",
-                    depthwise_initializer=tf.keras.initializers.constant(
-                        random_d_kernel
-                    ),
-                    pointwise_initializer=tf.keras.initializers.constant(
-                        random_p_kernel
-                    ),
+                    depthwise_initializer=const_init(random_d_kernel),
+                    pointwise_initializer=const_init(random_p_kernel),
                 ),
                 input_data=input_data,
                 should_run_eagerly=keras_should_run_eagerly,
             )
 
-        fp_output = testing_utils.layer_test(
-            layer,
-            kwargs=dict(
-                filters=3,
-                kernel_size=3,
-                depthwise_initializer=tf.keras.initializers.constant(
-                    np.sign(random_d_kernel)
-                ),
-                pointwise_initializer=tf.keras.initializers.constant(
-                    np.sign(random_p_kernel)
-                ),
-            ),
-            input_data=np.sign(input_data),
-            should_run_eagerly=keras_should_run_eagerly,
+        fp_model = tf.keras.models.Sequential(
+            [
+                layer(
+                    filters=3,
+                    kernel_size=3,
+                    depthwise_initializer=const_init(np.sign(random_d_kernel)),
+                    pointwise_initializer=const_init(np.sign(random_p_kernel)),
+                    input_shape=input_shape[1:],
+                )
+            ]
         )
 
-        np.testing.assert_allclose(quant_output, fp_output)
+        np.testing.assert_allclose(quant_output, fp_model.predict(np.sign(input_data)))
 
     def test_depthwise_layers(self, keras_should_run_eagerly):
         input_data = random_input((2, 3, 7, 6))
@@ -160,25 +151,23 @@ class TestLayers:
                     kernel_size=3,
                     depthwise_quantizer="ste_sign",
                     input_quantizer="ste_sign",
-                    depthwise_initializer=tf.keras.initializers.constant(random_weight),
+                    depthwise_initializer=const_init(random_weight),
                 ),
                 input_data=input_data,
[…]
