### 2019-11-13 bdc5037
Coalesce tensorflow imports
