### 2019-11-13 396d361
Simplify layer tests (#320)
* Simplify layer tests

This removes unnecessary computation from the layer tests which should result in a speed up.

* Coalesce tensorflow imports
