### 2019-11-13 10b1f81
Group and organize quantizer tests
