+++ larq/quantizers_test.py
@@ -6,289 +6,301 @@ import larq as lq
 from larq.testing_utils import generate_real_values_with_zeros
 
 
-@pytest.mark.parametrize("module", [lq.quantizers, tf.keras.activations])
-@pytest.mark.parametrize(
-    "name",
-    ["ste_sign", "approx_sign", "magnitude_aware_sign", "swish_sign", "ste_tern"],
-)
-def test_serialization(module, name):
-    fn = module.get(name)
-    ref_fn = getattr(lq.quantizers, name)
-    assert fn == ref_fn
-    assert type(fn.precision) == int
-    config = module.serialize(fn)
-    fn = module.deserialize(config)
-    assert fn == ref_fn
-    assert type(fn.precision) == int
-    fn = module.get(ref_fn)
-    assert fn == ref_fn
-    assert type(fn.precision) == int
+class TestCommonFunctionality:
+    """Test functionality common to all quantizers, like serialization and usage."""
 
-
-@pytest.mark.parametrize(
-    "ref_fn",
-    [
-        lq.quantizers.SteSign(),
-        lq.quantizers.SteHeaviside(),
-        lq.quantizers.MagnitudeAwareSign(),
-        lq.quantizers.SwishSign(),
-        lq.quantizers.SteTern(),
-    ],
-)
-def test_serialization_cls(ref_fn):
-    assert type(ref_fn.precision) == int
-    config = lq.quantizers.serialize(ref_fn)
-    fn = lq.quantizers.deserialize(config)
-    assert fn.__class__ == ref_fn.__class__
-
-
-def test_invalid_usage():
-    with pytest.raises(ValueError):
-        lq.quantizers.get(42)
-    with pytest.raises(ValueError):
-        lq.quantizers.get("unknown")
-
-
-@pytest.mark.parametrize(
-    "fn",
-    [
-        "ste_sign",
-        lq.quantizers.SteSign(),
-        "approx_sign",
-        "swish_sign",
-        lq.quantizers.SwishSign(),
-    ],
-)
-def test_xnor_binarization(fn):
-    x = tf.keras.backend.placeholder(ndim=2)
-    f = tf.keras.backend.function([x], [lq.quantizers.get(fn)(x)])
-    binarized_values = np.random.choice([-1, 1], size=(2, 5))
-    result = f([binarized_values])[0]
-    np.testing.assert_allclose(result, binarized_values)
-
-    real_values = generate_real_values_with_zeros()
-    result = f([real_values])[0]
-    assert not np.any(result == 0)
-    assert np.all(result[real_values < 0] == -1)
-    assert np.all(result[real_values >= 0] == 1)
-
-    zero_values = np.zeros((2, 5))
-    result = f([zero_values])[0]
-    assert np.all(result == 1)
-
-
-@pytest.mark.parametrize("fn", ["ste_heaviside", lq.quantizers.SteHeaviside()])
-def test_and_binarization(fn):
-    x = tf.keras.backend.placeholder(ndim=2)
-    f = tf.keras.backend.function([x], [lq.quantizers.get(fn)(x)])
-
-    binarized_values = np.random.choice([0, 1], size=(2, 5))
-    result = f([binarized_values])[0]
-    np.testing.assert_allclose(result, binarized_values)
-
-    real_values = generate_real_values_with_zeros()
-    result = f([real_values])[0]
-    assert np.all(result[real_values <= 0] == 0)
-    assert np.all(result[real_values > 0] == 1)
-
-
-@pytest.mark.parametrize(
-    "fn",
-    [
-        "ste_tern",
-        lq.quantizers.SteTern(),
-        lq.quantizers.SteTern(ternary_weight_networks=True),
-        lq.quantizers.SteTern(threshold_value=np.random.uniform(0.01, 0.8)),
-    ],
-)
[…]
