### 2019-11-13 272c1b0
Group and organize quantizer tests (#321)
* Group and organize quantizer tests

* Rename test to test_dorefa_ste_grad

* Fix pyflakes
