+++ larq/quantizers_test.py
@@ -6,289 +6,295 @@ import larq as lq
 from larq.testing_utils import generate_real_values_with_zeros
 
 
-@pytest.mark.parametrize("module", [lq.quantizers, tf.keras.activations])
-@pytest.mark.parametrize(
-    "name",
-    ["ste_sign", "approx_sign", "magnitude_aware_sign", "swish_sign", "ste_tern"],
-)
-def test_serialization(module, name):
-    fn = module.get(name)
-    ref_fn = getattr(lq.quantizers, name)
-    assert fn == ref_fn
-    assert type(fn.precision) == int
-    config = module.serialize(fn)
-    fn = module.deserialize(config)
-    assert fn == ref_fn
-    assert type(fn.precision) == int
-    fn = module.get(ref_fn)
-    assert fn == ref_fn
-    assert type(fn.precision) == int
+class TestCommonFunctionality:
+    """Test functionality common to all quantizers, like serialization and usage."""
+
+    @pytest.mark.parametrize("module", [lq.quantizers, tf.keras.activations])
+    @pytest.mark.parametrize(
+        "name",
+        ["ste_sign", "approx_sign", "magnitude_aware_sign", "swish_sign", "ste_tern"],
+    )
+    def test_serialization(self, module, name):
+        fn = module.get(name)
+        ref_fn = getattr(lq.quantizers, name)
+        assert fn == ref_fn
+        assert type(fn.precision) == int
+        config = module.serialize(fn)
+        fn = module.deserialize(config)
+        assert fn == ref_fn
+        assert type(fn.precision) == int
+        fn = module.get(ref_fn)
+        assert fn == ref_fn
+        assert type(fn.precision) == int
+
+    @pytest.mark.parametrize(
+        "ref_fn",
+        [
+            lq.quantizers.SteSign(),
+            lq.quantizers.SteHeaviside(),
+            lq.quantizers.MagnitudeAwareSign(),
+            lq.quantizers.SwishSign(),
+            lq.quantizers.SteTern(),
+        ],
+    )
+    def test_serialization_cls(self, ref_fn):
+        assert type(ref_fn.precision) == int
+        config = lq.quantizers.serialize(ref_fn)
+        fn = lq.quantizers.deserialize(config)
+        assert fn.__class__ == ref_fn.__class__
+
+    def test_invalid_usage(self):
+        with pytest.raises(ValueError):
+            lq.quantizers.get(42)
+        with pytest.raises(ValueError):
+            lq.quantizers.get("unknown")
 
 
-@pytest.mark.parametrize(
-    "ref_fn",
-    [
-        lq.quantizers.SteSign(),
-        lq.quantizers.SteHeaviside(),
-        lq.quantizers.MagnitudeAwareSign(),
-        lq.quantizers.SwishSign(),
-        lq.quantizers.SteTern(),
-    ],
-)
-def test_serialization_cls(ref_fn):
-    assert type(ref_fn.precision) == int
-    config = lq.quantizers.serialize(ref_fn)
-    fn = lq.quantizers.deserialize(config)
-    assert fn.__class__ == ref_fn.__class__
+class TestQuantization:
+    """Test binarization and ternarization."""
 
+    @pytest.mark.parametrize(
+        "fn",
+        [
+            "ste_sign",
+            lq.quantizers.SteSign(),
+            "approx_sign",
+            "swish_sign",
+            lq.quantizers.SwishSign(),
+        ],
+    )
+    def test_xnor_binarization(self, fn):
+        x = tf.keras.backend.placeholder(ndim=2)
[…]
