### 2019-11-14 1f9f70f
Move random_input() to testing_utils.py
