### 2019-11-14 aee8feb
Simplify testing_utils import in quantizers_test.py
