+++ larq/quantizers_test.py
@@ -3,7 +3,7 @@ import numpy as np
 
 import pytest
 import larq as lq
-from larq.testing_utils import generate_real_values_with_zeros
+from larq import testing_utils
 
 
 class TestCommonFunctionality:
@@ -70,7 +70,7 @@ class TestQuantization:
         result = f([binarized_values])[0]
         np.testing.assert_allclose(result, binarized_values)
 
-        real_values = generate_real_values_with_zeros()
+        real_values = testing_utils.generate_real_values_with_zeros()
         result = f([real_values])[0]
         assert not np.any(result == 0)
         assert np.all(result[real_values < 0] == -1)
@@ -89,7 +89,7 @@ class TestQuantization:
         result = f([binarized_values])[0]
         np.testing.assert_allclose(result, binarized_values)
 
-        real_values = generate_real_values_with_zeros()
+        real_values = testing_utils.generate_real_values_with_zeros()
         result = f([real_values])[0]
         assert np.all(result[real_values <= 0] == 0)
         assert np.all(result[real_values > 0] == 1)
@@ -132,7 +132,7 @@ class TestQuantization:
         assert np.any(result == 1)
         assert np.any(result == 0)
 
-        real_values = generate_real_values_with_zeros()
+        real_values = testing_utils.generate_real_values_with_zeros()
         result = f([real_values])[0]
         assert not np.any(result > 1)
         assert not np.any(result < -1)
@@ -146,7 +146,7 @@ class TestQuantization:
         test_threshold = 0.05  # This is the default
         f = tf.keras.backend.function([x], [lq.quantizers.get(fn)(x)])
 
-        real_values = generate_real_values_with_zeros()
+        real_values = testing_utils.generate_real_values_with_zeros()
         result = f([real_values])[0]
         assert np.all(result[real_values > test_threshold] == 1)
         assert np.all(result[real_values < -test_threshold] == -1)
@@ -160,7 +160,7 @@ class TestQuantization:
         fn = lq.quantizers.SteTern(threshold_value=test_threshold)
         f = tf.keras.backend.function([x], [fn(x)])
 
-        real_values = generate_real_values_with_zeros()
+        real_values = testing_utils.generate_real_values_with_zeros()
         result = f([real_values])[0]
         assert np.all(result[real_values > test_threshold] == 1)
         assert np.all(result[real_values < -test_threshold] == -1)
@@ -170,7 +170,7 @@ class TestQuantization:
 
     def test_ternarization_with_ternary_weight_networks(self):
         x = tf.keras.backend.placeholder(ndim=2)
-        real_values = generate_real_values_with_zeros()
+        real_values = testing_utils.generate_real_values_with_zeros()
         test_threshold = 0.7 * np.sum(np.abs(real_values)) / np.size(real_values)
         fn = lq.quantizers.SteTern(ternary_weight_networks=True)
         f = tf.keras.backend.function([x], [fn(x)])
@@ -188,7 +188,7 @@ class TestQuantization:
     def test_dorefa_quantize(self, fn):
         x = tf.keras.backend.placeholder(ndim=2)
         f = tf.keras.backend.function([x], [fn(x)])
-        real_values = generate_real_values_with_zeros()
+        real_values = testing_utils.generate_real_values_with_zeros()
         result = f([real_values])[0]
         k_bit = 2
         n = 2 ** k_bit - 1
@@ -212,7 +212,7 @@ class TestGradients:
         [lq.quantizers.ste_sign, lq.quantizers.ste_tern, lq.quantizers.ste_heaviside],
[…]
