### 2019-11-14 82afca6
Add test_layer_as_input_quantizer
