### 2019-11-14 14ca0f0
Add test_layer_as_kernel_quantizer
