### 2019-11-15 e93f2d7
Remove caplog length assert from test_separable_layer_warns
