### 2019-11-15 a76913a
Combine and parameterize layer as quantizer tests
