### 2019-11-15 2758d13
Run test_layer_as_quantizer in all keras modes
