+++ larq/quantizers_test.py
@@ -61,12 +61,12 @@ class TestCommonFunctionality:
             lq.quantizers.get("unknown")
 
     @pytest.mark.parametrize("quantizer", ["input_quantizer", "kernel_quantizer"])
-    def test_layer_as_quantizer(self, quantizer):
+    def test_layer_as_quantizer(self, quantizer, keras_should_run_eagerly):
         input_data = testing_utils.random_input((1, 10))
         model = tf.keras.Sequential(
             [lq.layers.QuantDense(1, **{quantizer: DummyTrainableQuantizer()})]
         )
-        model.compile(optimizer="sgd", loss="mse")  # TODO modes?
+        model.compile(optimizer="sgd", loss="mse", run_eagerly=keras_should_run_eagerly)
         model.fit(input_data, np.ones((1,)), epochs=1)
 
 
