### 2019-11-15 f44f187
Assert that dummy weight is trainable
