### 2019-11-15 6f18393
Quantizer layer tests (#322)
* Move random_input() to testing_utils.py

* Simplify testing_utils import in quantizers_test.py

* Add test_layer_as_input_quantizer

* Add test_layer_as_kernel_quantizer

* Remove caplog length assert from test_separable_layer_warns

* Combine and parameterize layer as quantizer tests

* Run test_layer_as_quantizer in all keras modes

* Assert that dummy weight is trainable
