+++ larq/quantizers_test.py
@@ -3,7 +3,18 @@ import numpy as np
 
 import pytest
 import larq as lq
-from larq.testing_utils import generate_real_values_with_zeros
+from larq import testing_utils
+
+
+class DummyTrainableQuantizer(tf.keras.layers.Layer):
+    """Used to test whether we can set layers as quantizers without any throws."""
+
+    def build(self, input_shape):
+        self.dummy_weight = self.add_weight("dummy_weight", trainable=True)
+        super().build(input_shape)
+
+    def call(self, inputs):
+        return self.dummy_weight * inputs
 
 
 class TestCommonFunctionality:
@@ -49,6 +60,20 @@ class TestCommonFunctionality:
         with pytest.raises(ValueError):
             lq.quantizers.get("unknown")
 
+    @pytest.mark.parametrize("quantizer", ["input_quantizer", "kernel_quantizer"])
+    def test_layer_as_quantizer(self, quantizer, keras_should_run_eagerly):
+        """Test whether a keras.layers.Layer can be used as quantizer."""
+
+        input_data = testing_utils.random_input((1, 10))
+
+        model = tf.keras.Sequential(
+            [lq.layers.QuantDense(1, **{quantizer: DummyTrainableQuantizer()})]
+        )
+        model.compile(optimizer="sgd", loss="mse", run_eagerly=keras_should_run_eagerly)
+        model.fit(input_data, np.ones((1,)), epochs=1)
+
+        assert any(["dummy_weight" in var.name for var in model.trainable_variables])
+
 
 class TestQuantization:
     """Test binarization and ternarization."""
@@ -70,7 +95,7 @@ class TestQuantization:
         result = f([binarized_values])[0]
         np.testing.assert_allclose(result, binarized_values)
 
-        real_values = generate_real_values_with_zeros()
+        real_values = testing_utils.generate_real_values_with_zeros()
         result = f([real_values])[0]
         assert not np.any(result == 0)
         assert np.all(result[real_values < 0] == -1)
@@ -89,7 +114,7 @@ class TestQuantization:
         result = f([binarized_values])[0]
         np.testing.assert_allclose(result, binarized_values)
 
-        real_values = generate_real_values_with_zeros()
+        real_values = testing_utils.generate_real_values_with_zeros()
         result = f([real_values])[0]
         assert np.all(result[real_values <= 0] == 0)
         assert np.all(result[real_values > 0] == 1)
@@ -132,7 +157,7 @@ class TestQuantization:
         assert np.any(result == 1)
         assert np.any(result == 0)
 
-        real_values = generate_real_values_with_zeros()
+        real_values = testing_utils.generate_real_values_with_zeros()
         result = f([real_values])[0]
         assert not np.any(result > 1)
         assert not np.any(result < -1)
@@ -146,7 +171,7 @@ class TestQuantization:
         test_threshold = 0.05  # This is the default
         f = tf.keras.backend.function([x], [lq.quantizers.get(fn)(x)])
 
-        real_values = generate_real_values_with_zeros()
+        real_values = testing_utils.generate_real_values_with_zeros()
         result = f([real_values])[0]
         assert np.all(result[real_values > test_threshold] == 1)
         assert np.all(result[real_values < -test_threshold] == -1)
@@ -160,7 +185,7 @@ class TestQuantization:
         fn = lq.quantizers.SteTern(threshold_value=test_threshold)
         f = tf.keras.backend.function([x], [fn(x)])
 
-        real_values = generate_real_values_with_zeros()
[…]
