### 2019-11-15 c532732
:arrow_up: 0.8.0 (#324)
