### 2019-11-15 007cd92
Simplify __repr__
