### 2019-11-15 e2a085f
Call delegated methods with *args, **kwargs
This should keep us compatible if TesnorFlow changes method arguments
