+++ larq/quantized_variable.py
@@ -54,22 +54,19 @@ class QuantizedVariable(tf.Variable):
         return self.latent_variable.read_value()
 
     @quantize
-    def sparse_read(self, indices, name=None):
-        """Reads the value of this variable sparsely, using `gather`."""
-        return self.latent_variable.sparse_read(indices, name=name)
+    def sparse_read(self, *args, **kwargs):
+        return self.latent_variable.sparse_read(*args, **kwargs)
 
     @quantize
-    def gather_nd(self, indices, name=None):
-        """Gather slices of the variable into a Tensor."""
-        return self.latent_variable.gather_nd(indices, name=name)
+    def gather_nd(self, *args, **kwargs):
+        return self.latent_variable.gather_nd(*args, **kwargs)
 
     def __getattr__(self, name):
         return getattr(self.latent_variable, name)
 
     @quantize
-    def _dense_var_to_tensor(self, dtype=None, name=None, as_ref=False):
-        """Converts this variable to a tensor."""
-        return self.latent_variable._dense_var_to_tensor(dtype, name, as_ref)
+    def _dense_var_to_tensor(self, *args, **kwargs):
+        return self.latent_variable._dense_var_to_tensor(*args, **kwargs)
 
     def eval(self, session=None):
         if self.quantizer:
@@ -117,8 +114,8 @@ class QuantizedVariable(tf.Variable):
     #     would be the same as the ref of the underlying variable, which would be
     #     strange as they are different Python objects.
 
-    def set_shape(self, shape):
-        return self.latent_variable.set_shape(self, shape)
+    def set_shape(self, *args, **kwargs):
+        return self.latent_variable.set_shape(self, *args, **kwargs)
 
     @property
     def trainable(self):
@@ -136,55 +133,53 @@ class QuantizedVariable(tf.Variable):
     def constraint(self):
         return self.latent_variable.constraint
 
-    def assign(self, value, use_locking=None, name=None, read_value=True):
-        return self.latent_variable.assign(value, use_locking, name, read_value)
+    def assign(self, *args, **kwargs):
+        return self.latent_variable.assign(*args, **kwargs)
 
-    def assign_add(self, delta, use_locking=None, name=None, read_value=True):
-        return self.latent_variable.assign_add(delta, use_locking, name, read_value)
+    def assign_add(self, *args, **kwargs):
+        return self.latent_variable.assign_add(*args, **kwargs)
 
-    def assign_sub(self, delta, use_locking=None, name=None, read_value=True):
-        return self.latent_variable.assign_sub(delta, use_locking, name, read_value)
+    def assign_sub(self, *args, **kwargs):
+        return self.latent_variable.assign_sub(*args, **kwargs)
 
-    def scatter_sub(self, sparse_delta, use_locking=False, name=None):
-        return self.latent_variable.scatter_sub(sparse_delta, use_locking, name)
+    def scatter_sub(self, *args, **kwargs):
+        return self.latent_variable.scatter_sub(*args, **kwargs)
 
-    def scatter_add(self, sparse_delta, use_locking=False, name=None):
-        return self.latent_variable.scatter_add(sparse_delta, use_locking, name)
+    def scatter_add(self, *args, **kwargs):
+        return self.latent_variable.scatter_add(*args, **kwargs)
 
-    def scatter_max(self, sparse_delta, use_locking=False, name=None):
-        return self.latent_variable.scatter_max(sparse_delta, use_locking, name)
+    def scatter_max(self, *args, **kwargs):
+        return self.latent_variable.scatter_max(*args, **kwargs)
 
-    def scatter_min(self, sparse_delta, use_locking=False, name=None):
[…]
