### 2019-11-15 a545369
Don't overwrite QuantizedDistributedVariable.__init__
This will not be needed in TF 2.1 anymore and is currently only used for saved models inside distribution strategies, so I think we should be save to remove it.

For more info see https://github.com/tensorflow/tensorflow/pull/34323
