### 2019-11-16 0e965b6
Bump pymdown-extensions from 6.1 to 6.2
Bumps [pymdown-extensions](https://github.com/facelessuser/pymdown-extensions) from 6.1 to 6.2.
- [Release notes](https://github.com/facelessuser/pymdown-extensions/releases)
- [Commits](https://github.com/facelessuser/pymdown-extensions/compare/6.1.0...6.2.0)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
