+++ setup.py
@@ -35,7 +35,7 @@ setup(
         "docs": [
             "mkdocs==1.0.4",
             "mkdocs-material==4.4.3",
-            "pymdown-extensions==6.1",
+            "pymdown-extensions==6.2",
             "mknotebooks==0.1.7",
             "mkdocs-minify-plugin==0.2.1",
             "larq-zoo==0.4.2",
