### 2019-11-16 #1746 @labarba
@lgeiger — Could we have a status update of your work on responding to reviewer comments? Let me know if you want us to set a reminder for you.
