### 2019-11-18 51aa649
Test correct inheritance in distributed case
