### 2019-11-18 9b6594b
Loose tf.Variable check to make TF 1.x tests pass
TODO: Check if this could lead to problems
