+++ larq/utils.py
@@ -1,4 +1,5 @@
 from contextlib import contextmanager
+
 from tensorflow.keras.utils import get_custom_objects
 
 
