+++ larq/models_test.py
@@ -1,8 +1,9 @@
+import numpy as np
 import pytest
 import tensorflow as tf
+
 import larq as lq
 from larq.models import ModelProfile
-import numpy as np
 
 
 def get_profile_model():
