+++ larq/constraints_test.py
@@ -1,6 +1,7 @@
-import pytest
 import numpy as np
+import pytest
 import tensorflow as tf
+
 import larq as lq
 from larq.testing_utils import generate_real_values_with_zeros
 
