### 2019-11-19 c4b77f3
Add Flake8 linting.
