+++ setup.cfg
@@ -2,7 +2,7 @@
 
 # Black compatibility.
 ignore = E203,E501,W503
-exclude = build,dist,env,.pytype
+exclude = build,dist,env,.pytype,larq/snapshots
 
 
 [isort]
@@ -22,14 +22,10 @@ skip =
 
 [pytype]
 
-inputs = .
+inputs = larq
 output = .pytype
-exclude =
-    build
-    dist
-    env
-    .pytype
-    examples
 # Keep going past errors to analyse as many files as possible.
 keep_going = True
+# Don't check use of imported modules, because we have no type-stubs for TF.
+strict_import = True
 python_version = 3.7
