### 2019-11-19 fe9ac22
Add type checking with pytype.
