### 2019-11-19 147ca25
Fix GitHub action triggers.
