+++ .github/workflows/unittest.yml
@@ -1,6 +1,10 @@
 name: Unittest
 
-on: [push]
+on:
+  push:
+    branches:
+      - master
+  pull_request: {}
 
 jobs:
   tests:
