+++ .github/workflows/lint.yml
@@ -1,6 +1,10 @@
 name: Lint
 
-on: [push]
+on:
+  push:
+    branches:
+      - master
+  pull_request: {}
 
 jobs:
   build:
