+++ .github/workflows/lint.yml
@@ -24,5 +24,7 @@ jobs:
         run: flake8
       - name: Black code style
         run: black --check --target-version py36 --exclude 'build/|\.mypy_cache/|\.venv/|env/|larq/snapshots/|.pytype/' .
+      - name: Check import order with isort
+        run: isort --check-only --diff
       - name: Type check with PyType
         run: pytype
