+++ .github/workflows/lint.yml
@@ -7,7 +7,7 @@ on:
   pull_request: {}
 
 jobs:
-  build:
+  lint:
     runs-on: ubuntu-latest
 
     steps:
