### 2019-11-19 14c4bbe
Second paper draft [skip ci]
