+++ paper/paper.bib
@@ -11,7 +11,13 @@
     year         = 2015,
     howpublished = {\url{https://keras.io}}
 }
-@article{2019arXiv190602107H,
+@inproceedings{pytorch,
+  title={Automatic Differentiation in {PyTorch}},
+  author={Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory and Yang, Edward and DeVito, Zachary and Lin, Zeming and Desmaison, Alban and Antiga, Luca and Lerer, Adam},
+  booktitle={NIPS Autodiff Workshop},
+  year={2017}
+}
+@article{bop,
     title        = {Latent Weights Do Not Exist: Rethinking Binarized Neural Network Optimization},
     author       = {{Helwegen}, Koen and {Widdicombe}, James and {Geiger}, Lukas and {Liu}, Zechun and {Cheng}, Kwang-Ting and {Nusselder}, Roeland},
     year         = 2019,
@@ -27,7 +33,7 @@
     adsurl       = {https://ui.adsabs.harvard.edu/abs/2019arXiv190602107H},
     adsnote      = {Provided by the SAO/NASA Astrophysics Data System}
 }
-@incollection{NIPS2016_6573,
+@incollection{binarynet,
     title        = {Binarized Neural Networks},
     author       = {Hubara, Itay and Courbariaux, Matthieu and Soudry, Daniel and El-Yaniv, Ran and Bengio, Yoshua},
     booktitle    = {Advances in Neural Information Processing Systems 29},
@@ -36,4 +42,56 @@
     year         = {2016},
     publisher    = {Curran Associates, Inc.},
     url          = {http://papers.nips.cc/paper/6573-binarized-neural-networks.pdf}
+}
+@article{binary_dense_net,
+  author    = {Joseph Bethge and
+               Haojin Yang and
+               Marvin Bornstein and
+               Christoph Meinel},
+  title     = {Back to Simplicity: How to Train Accurate BNNs from Scratch?},
+  journal   = {CoRR},
+  volume    = {abs/1906.08637},
+  year      = {2019},
+  url       = {http://arxiv.org/abs/1906.08637},
+  archivePrefix = {arXiv},
+  eprint    = {1906.08637},
+  timestamp = {Mon, 24 Jun 2019 17:28:45 +0200},
+  biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1906-08637},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+@article{bireal_net,
+  author    = {Zechun Liu and
+               Baoyuan Wu and
+               Wenhan Luo and
+               Xin Yang and
+               Wei Liu and
+               Kwang{-}Ting Cheng},
+  title     = {Bi-Real Net: Enhancing the Performance of 1-bit CNNs With Improved
+               Representational Capability and Advanced Training Algorithm},
+  journal   = {CoRR},
+  volume    = {abs/1808.00278},
+  year      = {2018},
+  url       = {http://arxiv.org/abs/1808.00278},
+  archivePrefix = {arXiv},
+  eprint    = {1808.00278},
+  timestamp = {Thu, 12 Sep 2019 15:28:48 +0200},
+  biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1808-00278},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+@article{xnor_net,
+  author    = {Mohammad Rastegari and
+               Vicente Ordonez and
+               Joseph Redmon and
+               Ali Farhadi},
+  title     = {XNOR-Net: ImageNet Classification Using Binary Convolutional Neural
+               Networks},
[…]
