### 2019-11-19 2b262bf
Improve linting (#328)
* Sort imports.

* Add Flake8 linting.

* Add type checking with pytype.

* Fix GitHub action triggers.

* Check import sorting.

* Rename linting.

* Update test dependencies.
