+++ larq/layers_test.py
@@ -1,10 +1,11 @@
-import tensorflow as tf
-import numpy as np
-import larq as lq
-import pytest
 import inspect
-from larq import testing_utils
 
+import numpy as np
+import pytest
+import tensorflow as tf
+
+import larq as lq
+from larq import testing_utils
 
 PARAMS_ALL_LAYERS = [
     (lq.layers.QuantDense, tf.keras.layers.Dense, (3, 2), dict(units=3)),
@@ -273,9 +274,9 @@ def test_layer_kwargs(quant_layer, layer):
     ):
         try:
             quant_params_list.remove(p)
-        except:
+        except ValueError:
             pass
     assert quant_params_list == params_list
 
     for param in params_list:
-        assert quant_params.get(param).default == params.get(param).default
+        assert quant_params.get(param).default == params.get(param).default  # type: ignore
