+++ larq/__init__.py
@@ -1,12 +1,12 @@
-import larq.layers as layers
 import larq.activations as activations
 import larq.callbacks as callbacks
 import larq.constraints as constraints
+import larq.layers as layers
 import larq.math as math
 import larq.metrics as metrics
 import larq.models as models
-import larq.quantizers as quantizers
 import larq.optimizers as optimizers
+import larq.quantizers as quantizers
 
 __all__ = [
     "layers",
