### 2019-11-20 6ccaf1d
Fix CaseOptimizer usage example in docs
