### 2019-11-20 ec3c2a1
Fix other CaseOptimizer doc
