### 2019-11-20 a0a42d3
Fix model summary
