### 2019-11-20 2104b18
Why don't you like my code flake8
Well I guess it makes sense to now catch SystemExit or Keyboard interrupts...
