+++ larq/models.py
@@ -83,7 +83,7 @@ def _format_table_entry(x, units=1):
         if type(x) == str or x == 0 or units == 1:
             return x
         return x / units
-    except:
+    except Exception:
         return "?"
 
 
