### 2019-11-20 413008f
Fix examples in documentation for CaseOptimizer and Bop (#331)
* Fix CaseOptimizer usage example in docs

* Fix other CaseOptimizer doc
