### 2019-11-20 0fa974f
Add references to dabnn and BMX-Net
