+++ paper/paper.bib
@@ -17,6 +17,20 @@
   booktitle={NIPS Autodiff Workshop},
   year={2017}
 }
+@article{bmxnetv2,
+  title = {Training Competitive Binary Neural Networks from Scratch},
+  author = {Joseph Bethge and Marvin Bornstein and Adrian Loy and Haojin Yang and Christoph Meinel},
+  journal = {ArXiv e-prints},
+  archivePrefix = "arXiv",
+  eprint = {1812.01965},
+  Year = {2018}
+}
+@misc{dabnn,
+  Author = {Jianhao Zhang and Yingwei Pan and Ting Yao and He Zhao and Tao Mei},
+  Title = {daBNN: A Super Fast Inference Framework for Binary Neural Networks on ARM devices},
+  Year = {2019},
+  Eprint = {arXiv:1908.05858},
+}
 @article{bop,
     title        = {Latent Weights Do Not Exist: Rethinking Binarized Neural Network Optimization},
     author       = {{Helwegen}, Koen and {Widdicombe}, James and {Geiger}, Lukas and {Liu}, Zechun and {Cheng}, Kwang-Ting and {Nusselder}, Roeland},
