### 2019-11-20 50cb5c7
Fix try except in model summary (#332)
