### 2019-11-20 06150e3
Expand description of larq/zoo
