+++ paper/paper.bib
@@ -7,7 +7,7 @@
 }
 @misc{chollet2015keras,
     title        = {Keras},
-    author       = {Chollet, Fran\c{c}ois and others},
+    author       = {Chollet, Fran\c{c}ois},
     year         = 2015,
     howpublished = {\url{https://keras.io}}
 }
@@ -107,5 +107,24 @@
   eprint    = {1603.05279},
   timestamp = {Mon, 13 Aug 2018 16:47:22 +0200},
   biburl    = {https://dblp.org/rec/bib/journals/corr/RastegariORF16},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+@article{dorefa,
+  author    = {Shuchang Zhou and
+               Zekun Ni and
+               Xinyu Zhou and
+               He Wen and
+               Yuxin Wu and
+               Yuheng Zou},
+  title     = {DoReFa-Net: Training Low Bitwidth Convolutional Neural Networks with
+               Low Bitwidth Gradients},
+  journal   = {CoRR},
+  volume    = {abs/1606.06160},
+  year      = {2016},
+  url       = {http://arxiv.org/abs/1606.06160},
+  archivePrefix = {arXiv},
+  eprint    = {1606.06160},
+  timestamp = {Mon, 13 Aug 2018 16:48:10 +0200},
+  biburl    = {https://dblp.org/rec/bib/journals/corr/ZhouNZWWZ16},
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
