### 2019-11-21 974fb27
Bump matplotlib from 3.1.1 to 3.1.2
Bumps [matplotlib](https://github.com/matplotlib/matplotlib) from 3.1.1 to 3.1.2.
- [Release notes](https://github.com/matplotlib/matplotlib/releases)
- [Commits](https://github.com/matplotlib/matplotlib/compare/v3.1.1...v3.1.2)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
