+++ requirements.txt
@@ -2,5 +2,5 @@ tensorflow==2.0.0
 tensorflow_datasets==1.3.0
 larq_zoo==0.4.2
 Pillow==6.2.1
-matplotlib==3.1.1
+matplotlib==3.1.2
 gast==0.2.2  # See https://github.com/tensorflow/tensorflow/issues/32319
