+++ paper/paper.bib
@@ -12,24 +12,24 @@
     howpublished = {\url{https://keras.io}}
 }
 @inproceedings{pytorch,
-  title={Automatic Differentiation in {PyTorch}},
-  author={Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory and Yang, Edward and DeVito, Zachary and Lin, Zeming and Desmaison, Alban and Antiga, Luca and Lerer, Adam},
-  booktitle={NIPS Autodiff Workshop},
-  year={2017}
+    title        = {Automatic Differentiation in {PyTorch}},
+    author       = {Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory and Yang, Edward and DeVito, Zachary and Lin, Zeming and Desmaison, Alban and Antiga, Luca and Lerer, Adam},
+    year         = 2017,
+    booktitle    = {NIPS Autodiff Workshop}
 }
 @article{bmxnetv2,
-  title = {Training Competitive Binary Neural Networks from Scratch},
-  author = {Joseph Bethge and Marvin Bornstein and Adrian Loy and Haojin Yang and Christoph Meinel},
-  journal = {ArXiv e-prints},
-  archivePrefix = "arXiv",
-  eprint = {1812.01965},
-  Year = {2018}
+    title        = {Training Competitive Binary Neural Networks from Scratch},
+    author       = {Joseph Bethge and Marvin Bornstein and Adrian Loy and Haojin Yang and Christoph Meinel},
+    year         = 2018,
+    journal      = {ArXiv e-prints},
+    archiveprefix = {arXiv},
+    eprint       = {1812.01965}
 }
 @misc{dabnn,
-  Author = {Jianhao Zhang and Yingwei Pan and Ting Yao and He Zhao and Tao Mei},
-  Title = {daBNN: A Super Fast Inference Framework for Binary Neural Networks on ARM devices},
-  Year = {2019},
-  Eprint = {arXiv:1908.05858},
+    title        = {daBNN: A Super Fast Inference Framework for Binary Neural Networks on ARM devices},
+    author       = {Jianhao Zhang and Yingwei Pan and Ting Yao and He Zhao and Tao Mei},
+    year         = 2019,
+    eprint       = {arXiv:1908.05858}
 }
 @article{bop,
     title        = {Latent Weights Do Not Exist: Rethinking Binarized Neural Network Optimization},
@@ -38,93 +38,124 @@
     month        = jun,
     journal      = {arXiv e-prints},
     pages        = {arXiv:1906.02107},
+    url          = {https://arxiv.org/abs/1906.02107},
     keywords     = {Computer Science - Machine Learning, Statistics - Machine Learning},
     eid          = {arXiv:1906.02107},
     archiveprefix = {arXiv},
     eprint       = {1906.02107},
     primaryclass = {cs.LG},
-    url          = {https://arxiv.org/abs/1906.02107},
     adsurl       = {https://ui.adsabs.harvard.edu/abs/2019arXiv190602107H},
     adsnote      = {Provided by the SAO/NASA Astrophysics Data System}
 }
 @incollection{binarynet,
     title        = {Binarized Neural Networks},
     author       = {Hubara, Itay and Courbariaux, Matthieu and Soudry, Daniel and El-Yaniv, Ran and Bengio, Yoshua},
+    year         = 2016,
     booktitle    = {Advances in Neural Information Processing Systems 29},
-    editor       = {D. D. Lee and M. Sugiyama and U. V. Luxburg and I. Guyon and R. Garnett},
-    pages        = {4107--4115},
-    year         = {2016},
     publisher    = {Curran Associates, Inc.},
[…]
