### 2019-11-22 80f0853
Fix reloading of Bop when used with schedules (#334)
Co-Authored-By: James Widdicombe <j.y.widdicombe@gmail.com>
