+++ larq/optimizers_test.py
@@ -151,3 +151,14 @@ class TestBopOptimizer:
                 default_optimizer=tf.keras.optimizers.Adam(0.01),
             ),
         )
+
+    @pytest.mark.parametrize(
+        "hyper", [5e-4, tf.keras.optimizers.schedules.PolynomialDecay(5e-4, 100)],
+    )
+    def test_bop_serialization_schedule(self, hyper):
+        bop = lq.optimizers.Bop(gamma=hyper, threshold=hyper,)
+        new_bop = lq.optimizers.Bop.from_config(bop.get_config())
+        assert isinstance(new_bop._get_hyper("gamma"), type(bop._get_hyper("gamma")))
+        assert isinstance(
+            new_bop._get_hyper("threshold"), type(bop._get_hyper("threshold"))
+        )
