### 2019-11-22 176b258
Update setup.py (#335)
