+++ paper/paper.bib
@@ -1,11 +1,11 @@
-@misc{tensorflow2015-whitepaper,
+@misc{tensorflow,
     title        = {{TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
     author       = {Mart\'{\i}n~Abadi and Ashish~Agarwal and Paul~Barham and Eugene~Brevdo and Zhifeng~Chen and Craig~Citro and Greg~S.~Corrado and Andy~Davis and Jeffrey~Dean and Matthieu~Devin and Sanjay~Ghemawat and Ian~Goodfellow and Andrew~Harp and Geoffrey~Irving and Michael~Isard and Yangqing Jia and Rafal~Jozefowicz and Lukasz~Kaiser and Manjunath~Kudlur and Josh~Levenberg and Dandelion~Man\'{e} and Rajat~Monga and Sherry~Moore and Derek~Murray and Chris~Olah and Mike~Schuster and Jonathon~Shlens and Benoit~Steiner and Ilya~Sutskever and Kunal~Talwar and Paul~Tucker and Vincent~Vanhoucke and Vijay~Vasudevan and Fernanda~Vi\'{e}gas and Oriol~Vinyals and Pete~Warden and Martin~Wattenberg and Martin~Wicke and Yuan~Yu and Xiaoqiang~Zheng},
     year         = 2015,
     url          = {https://www.tensorflow.org/},
     note         = {Software available from tensorflow.org}
 }
-@misc{chollet2015keras,
+@misc{keras,
     title        = {Keras},
     author       = {Chollet, Fran\c{c}ois},
     year         = 2015,
@@ -122,7 +122,7 @@
     issn         = 280836,
     pmid         = 134
 }
-@article{hintoncoursera,
+@article{hinton_coursera,
     title        = {Neural networks for machine learning},
     author       = {Hinton, Geoffrey},
     year         = 2012,
