### 2019-11-22 751bff2
Add SGD citation
