+++ paper/paper.bib
@@ -159,3 +159,12 @@
     issn         = 154040,
     url          = {http://arxiv.org/abs/1811.10413}
 }
+@article{bottou1991,
+    title        = {Stochastic gradient learning in neural networks},
+    author       = {Bottou, L{\'e}on},
+    journal      = {Proceedings of Neuro Nimes},
+    volume       = {91},
+    number       = {8},
+    pages        = {12},
+    year         = {1991}
+}
