### 2019-11-22 1e17088
Better description of BNNs
