+++ paper/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 Bringing the power of deep learning outside of data centers can transform society: self-driving cars, mobile-based neural networks, and autonomous drones all have the potential to revolutionize everyday lives.
 However, existing neural networks that use 32 bits to encode each weight and activation have an energy budget which is far beyond the scope of many of these applications. One common way to improve computational efficiency is to reduce the precision of the network to 16-bit or 8-bit, also known as quantization.
-Binarized Neural Networks (BNNs) represent and extreme case of quantization. In these networks both weights and activations are restricted to $\{-1, +1\}$ [@binarynet]. Compared to an equivalent 8-bit quantized network BNNs require 8 times less memory and can lead to a theoretical 64 times speedup when deployed on optimized hardware.
+Binarized Neural Networks (BNNs) represent an extreme case of quantized networks, that cannot be viewed as approximations to real valued networks and therefore require special tools and optimization strategies [@bop]. In these networks both weights and activations are restricted to $\{-1, +1\}$ [@binarynet]. Compared to an equivalent 8-bit quantized network BNNs require 8 times less memory and can lead to a theoretical 64 times speedup when deployed on optimized hardware.
 However, many open research questions remain until the widespread use of BNNs and other extremely quantized neural networks in industry. [`larq`](https://larq.dev) is an ecosystem of Python packages for BNNs and other Quantized Neural Networks (QNNs). It is intended to facilitate research to resolve these outstanding questions.
 
 # Background: Neural Network Binarization
