### 2019-11-22 c29bde7
Apply suggestions from code review
Co-Authored-By: Koen Helwegen <koen@plumerai.com>
