+++ paper/paper.bib
@@ -74,7 +74,7 @@
     title        = {Bi-Real Net: Enhancing the Performance of 1-bit CNNs With Improved Representational Capability and Advanced Training Algorithm},
     author       = {Zechun Liu and Baoyuan Wu and Wenhan Luo and Xin Yang and Wei Liu and Kwang{-}Ting Cheng},
     year         = 2018,
-    journal      = {CoRR},
+    journal      = {ECCV},
     volume       = {abs/1808.00278},
     url          = {http://arxiv.org/abs/1808.00278},
     archiveprefix = {arXiv},
