### 2019-11-22 f80db7e
Second paper draft (#329)
* Second paper draft [skip ci]

* Add references to dabnn and BMX-Net

* Expand description of larq/zoo

* Add BNN Background sections

* Fix references

* Add SGD citation

* Minor layout changes

* Better description of BNNs

* Apply suggestions from code review

Co-Authored-By: Koen Helwegen <koen@plumerai.com>
