+++ paper/paper.bib
@@ -1,39 +1,170 @@
-@misc{tensorflow2015-whitepaper,
+@misc{tensorflow,
     title        = {{TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
     author       = {Mart\'{\i}n~Abadi and Ashish~Agarwal and Paul~Barham and Eugene~Brevdo and Zhifeng~Chen and Craig~Citro and Greg~S.~Corrado and Andy~Davis and Jeffrey~Dean and Matthieu~Devin and Sanjay~Ghemawat and Ian~Goodfellow and Andrew~Harp and Geoffrey~Irving and Michael~Isard and Yangqing Jia and Rafal~Jozefowicz and Lukasz~Kaiser and Manjunath~Kudlur and Josh~Levenberg and Dandelion~Man\'{e} and Rajat~Monga and Sherry~Moore and Derek~Murray and Chris~Olah and Mike~Schuster and Jonathon~Shlens and Benoit~Steiner and Ilya~Sutskever and Kunal~Talwar and Paul~Tucker and Vincent~Vanhoucke and Vijay~Vasudevan and Fernanda~Vi\'{e}gas and Oriol~Vinyals and Pete~Warden and Martin~Wattenberg and Martin~Wicke and Yuan~Yu and Xiaoqiang~Zheng},
     year         = 2015,
     url          = {https://www.tensorflow.org/},
     note         = {Software available from tensorflow.org}
 }
-@misc{chollet2015keras,
+@misc{keras,
     title        = {Keras},
-    author       = {Chollet, Fran\c{c}ois and others},
+    author       = {Chollet, Fran\c{c}ois},
     year         = 2015,
     howpublished = {\url{https://keras.io}}
 }
-@article{2019arXiv190602107H,
+@inproceedings{pytorch,
+    title        = {Automatic Differentiation in {PyTorch}},
+    author       = {Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory and Yang, Edward and DeVito, Zachary and Lin, Zeming and Desmaison, Alban and Antiga, Luca and Lerer, Adam},
+    year         = 2017,
+    booktitle    = {NIPS Autodiff Workshop}
+}
+@article{bmxnetv2,
+    title        = {Training Competitive Binary Neural Networks from Scratch},
+    author       = {Joseph Bethge and Marvin Bornstein and Adrian Loy and Haojin Yang and Christoph Meinel},
+    year         = 2018,
+    journal      = {ArXiv e-prints},
+    archiveprefix = {arXiv},
+    eprint       = {1812.01965}
+}
+@misc{dabnn,
+    title        = {daBNN: A Super Fast Inference Framework for Binary Neural Networks on ARM devices},
+    author       = {Jianhao Zhang and Yingwei Pan and Ting Yao and He Zhao and Tao Mei},
+    year         = 2019,
+    eprint       = {arXiv:1908.05858}
+}
+@article{bop,
     title        = {Latent Weights Do Not Exist: Rethinking Binarized Neural Network Optimization},
     author       = {{Helwegen}, Koen and {Widdicombe}, James and {Geiger}, Lukas and {Liu}, Zechun and {Cheng}, Kwang-Ting and {Nusselder}, Roeland},
     year         = 2019,
     month        = jun,
     journal      = {arXiv e-prints},
     pages        = {arXiv:1906.02107},
+    url          = {https://arxiv.org/abs/1906.02107},
     keywords     = {Computer Science - Machine Learning, Statistics - Machine Learning},
     eid          = {arXiv:1906.02107},
     archiveprefix = {arXiv},
     eprint       = {1906.02107},
     primaryclass = {cs.LG},
[…]
