### 2019-11-22 0a6be9d
Don't overwrite add_weight
This is so QuantizedVariable will wrap AutoCastVariable and not vice versa in order to calculate quantization in mixed precision.
