### 2019-11-23 2431c9c
Add explicit quantized scope
This is necessary so we don't have to worry about properly supporting latent weight regulatization and saving
