### 2019-11-23 90e439d
Make quantized scope thread local
