### 2019-11-23 12f48cd
Remove outdated comment
