+++ larq/quantized_variable.py
@@ -107,8 +107,6 @@ class QuantizedVariable(tf.Variable):
     #
     # We do not define the following methods from Variable for the following
     # reasons:
-    #   * 'count_up_to': This method only applies to int variables, which cannot
-    #     be wrapped with an QuantizedVariable.
     #   * 'experimental_ref': Instead we inherit the definition from Variable.
     #     If we defined and delegated to Variable, the ref of an QuantizedVariable
     #     would be the same as the ref of the underlying variable, which would be
