### 2019-11-23 329ab12
Add docstrings for quantized scope
