### 2019-11-23 590c6a0
Respect quantized scope in QuantizedVariable.eval
