### 2019-11-23 1867caa
Simplify _add_variable_with_custom_getter
