### 2019-11-23 e4a592c
Silence pytype
It looks like it doesn't liek DistributedVariable
