+++ larq/quantized_variable_test.py
@@ -8,4 +8,4 @@ def test_inheritance(distribute_scope):
     variable = tf.Variable(3.0)
     quantized_variable = create_quantized_variable(variable)
     assert isinstance(quantized_variable, tf.Variable)
-    assert isinstance(quantized_variable, DistributedVariable) is distribute_scope
+    assert isinstance(quantized_variable, DistributedVariable) is distribute_scope  # type: ignore
