### 2019-11-25 8cd9b6f
Set default session in graph_mode fixture
