+++ larq/conftest_test.py
@@ -2,7 +2,11 @@ import tensorflow as tf
 
 
 def test_eager_and_graph_mode_fixture(eager_and_graph_mode):
-    assert tf.executing_eagerly() == (eager_and_graph_mode == "eager")
+    if eager_and_graph_mode == "eager":
+        assert tf.executing_eagerly()
+    else:
+        assert not tf.executing_eagerly()
+        assert tf.compat.v1.get_default_session() is not None
 
 
 def test_eager_mode_fixture(eager_mode):
@@ -11,6 +15,7 @@ def test_eager_mode_fixture(eager_mode):
 
 def test_graph_mode_fixture(graph_mode):
     assert not tf.executing_eagerly()
+    assert tf.compat.v1.get_default_session() is not None
 
 
 def test_distribute_scope(distribute_scope):
