### 2019-11-25 c43dd2c
Add eval helper to write graph and eager compatible tests
