### 2019-11-25 #1746 @terrytangyuan
@sbrugman @EduPH @casperdcl Could you take another look at the updated paper when you get a chance?
