### 2019-11-25 486106a
Test operator overloads
