### 2019-11-25 eb8bd51
Overwrite .numpy() and test equality
