### 2019-11-25 039e088
Only allow resource variables for now
AutocastVariable does this too
