### 2019-11-25 7810674
Fix TF 1.14 distributed test pass
