### 2019-11-27 a02452f
Bump altair from 3.2.0 to 3.3.0 (#337)
Bumps [altair](https://github.com/altair-viz/altair) from 3.2.0 to 3.3.0.
- [Release notes](https://github.com/altair-viz/altair/releases)
- [Changelog](https://github.com/altair-viz/altair/blob/master/CHANGES.md)
- [Commits](https://github.com/altair-viz/altair/compare/v3.2.0...v3.3.0)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
