### 2019-11-28 5706879
Update pytype requirement from ~=2019.10.17 to >=2019.10.17,<2019.12.0 (#338)
Updates the requirements on [pytype](https://github.com/google/pytype) to permit the latest version.
- [Release notes](https://github.com/google/pytype/releases)
- [Changelog](https://github.com/google/pytype/blob/master/CHANGELOG)
- [Commits](https://github.com/google/pytype/compare/2019.10.17...2019.11.27)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
