+++ setup.py
@@ -33,7 +33,7 @@ setup(
             "pytest>=5.2.4,<5.4.0",
             "pytest-cov~=2.8.1",
             "pytest-xdist~=1.30.0",
-            "pytype~=2019.10.17",
+            "pytype>=2019.10.17,<2019.12.0",
             "snapshottest~=0.5.1",
         ],
         "docs": [
