+++ .github/workflows/unittest.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        tf-version: [1.14.0, 1.15.0, 2.0.0]
+        tf-version: [1.14.0, 1.15.0, 2.0.0, 2.1.0rc0]
 
     steps:
       - uses: actions/checkout@v1
