### 2019-11-28 64def21
Make warning test less fragile
Co-Authored-By: Leon Overweel <l.overweel@gmail.com>
