### 2019-11-28 8ee9d05
Fix reloading of case optimizer weights (#340)
* Fix reloading of case optimizer weights

Co-Authored-By: Leon Overweel <l.overweel@gmail.com>

* Make warning test less fragile

Co-Authored-By: Leon Overweel <l.overweel@gmail.com>
