### 2019-11-28 a337b83
Bump larq-zoo from 0.4.2 to 0.5.0 (#342)
Bumps [larq-zoo](https://github.com/plumerai/larq-zoo) from 0.4.2 to 0.5.0.
- [Release notes](https://github.com/plumerai/larq-zoo/releases)
- [Commits](https://github.com/plumerai/larq-zoo/compare/v0.4.2...v0.5.0)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
