### 2019-11-29 ff33566
Simplify handling of quantization
