### 2019-11-29 d88b1d9
Return new instances of Quantized variable after assignment
