### 2019-12-03 5a5e8c5
Pass quantizers and precision to create_quantized_variable
