### 2019-12-03 1bc0dd2
Use create_quantized_variable to create new instance
See https://github.com/tensorflow/tensorflow/pull/34779/files/147de48ad973a6a05e8113af815988014652caf2#diff-8db8dd75efb914430be51c632dfa452e
