### 2019-12-03 #1746 @EduPH
The API and the paper have improved a lot since the beginning of the review. Just some typos or minor corrections in the paper:

real valued -> real-valued
networks in industry -> networks in the industry
facilitate research to resolve -> facilitate researchers to resolve
libaries -> libraries
commmon -> common
though in absence of a commmon API> though, in absence of a common API
