### 2019-12-03 876c878
Fix typos in paper
