### 2019-12-03 f7ec07a
Update paper/paper.md
Co-Authored-By: Adam Hillier <7688302+AdamHillier@users.noreply.github.com>
