### 2019-12-03 3298637
Fix typos in paper (#345)
* Fix typos in paper

* Update paper/paper.md

Co-Authored-By: Adam Hillier <7688302+AdamHillier@users.noreply.github.com>
