### 2019-12-03 #1746 @lgeiger
> Just some typos or minor corrections in the paper

Thanks a lot for the corrections, I updated the paper :+1:
