### 2019-12-04 924aaa8
Only wrap resource variable in QuantizedVariable
See https://github.com/tensorflow/tensorflow/pull/34779#discussion_r353481035
