### 2019-12-05 c97f75d
Add _maybe_wrap function to simplify wrapping of QuantizedVariable
