### 2019-12-06 201af9a
Pass *args, **kwargs to DistributedVariable.get
