### 2019-12-06 1614e5a
Fix threading for quantization in distributed training
