+++ requirements.txt
@@ -1,5 +1,5 @@
 tensorflow==2.0.0
-tensorflow_datasets==1.3.0
+tensorflow_datasets==1.3.1
 larq_zoo==0.5.0
 Pillow==6.2.1
 matplotlib==3.1.2
