### 2019-12-06 0f9f752
Update pytype requirement (#350)
Updates the requirements on [pytype](https://github.com/google/pytype) to permit the latest version.
- [Release notes](https://github.com/google/pytype/releases)
- [Changelog](https://github.com/google/pytype/blob/master/CHANGELOG)
- [Commits](https://github.com/google/pytype/compare/2019.10.17...2019.12.06)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
