### 2019-12-08 922152b
Improve __repr__
