+++ larq/quantized_variable.py
@@ -71,18 +71,17 @@ class QuantizedVariable(tf.Variable):
         pass
 
     def __repr__(self):
-        if tf.executing_eagerly() and not self._in_graph_mode:
-            return (
-                f"<{self.__class__.__name__} '{self.name}' shape={self.shape} "
-                f"dtype={self.dtype.name} quantizer={self.quantizer.__repr__()} "
-                f"precision={self.precision} "
-                f"numpy={ops.numpy_text(self.read_value(), is_repr=True)}>"
-            )
-        return (
-            f"<{self.__class__.__name__} '{self.name}' shape={self.shape} "
-            f"dtype={self.dtype.name} quantizer={self.quantizer.__repr__()} "
-            f"precision={self.precision}>"
+        repr_ = (
+            f"<{self.__class__.__name__} '{self.name}' "
+            f"shape={self.shape} dtype={self.dtype.name}"
         )
+        if self.quantizer is not None:
+            repr_ += f" quantizer={_get_name(self.quantizer)}"
+        if self.precision is not None:
+            repr_ += f" precision={self.precision}"
+        if tf.executing_eagerly() and not self._in_graph_mode:
+            return f"{repr_} numpy={ops.numpy_text(self.read_value(), is_repr=True)}>"
+        return f"{repr_}>"
 
     # Method delegations: We delegate the following methods to self.latent_variable.
     # Each of these methods simply calls the same method on self.latent_variable. The
@@ -289,3 +288,10 @@ def _maybe_wrap(variable, quantizer, precision, wrap=True):
     if wrap and resource_variable_ops.is_resource_variable(variable):
         return create_quantized_variable(variable, quantizer, precision)
     return variable
+
+
+def _get_name(obj):
+    try:
+        return obj.__name__
+    except AttributeError:
+        return obj.__class__.__name__
