### 2019-12-08 132680c
Quantize _as_graph_element
