### 2019-12-08 dbad0d0
Add unittests for QuantizedVariable
