+++ larq/quantized_variable_test.py
@@ -1,10 +1,10 @@
-import numpy as np
 import pytest
 import tensorflow as tf
+from numpy.testing import assert_almost_equal, assert_array_equal
 from tensorflow.python.distribute.values import DistributedVariable
 
 from larq import quantized_scope
-from larq.quantized_variable import create_quantized_variable
+from larq.quantized_variable import QuantizedVariable, create_quantized_variable
 from larq.testing_utils import evaluate
 
 
@@ -26,8 +26,104 @@ def test_quantize_scope(quantized):
 def test_inheritance(distribute_scope):
     variable = get_var(3.0)
     quantized_variable = create_quantized_variable(variable)
+    assert isinstance(quantized_variable, QuantizedVariable)
     assert isinstance(quantized_variable, tf.Variable)
     assert isinstance(quantized_variable, DistributedVariable) is distribute_scope  # type: ignore
+
+
+def test_read(distribute_scope, eager_and_graph_mode):
+    x = create_quantized_variable(get_var(3.5), quantizer=lambda x: 2 * x)
+    evaluate(x.initializer)
+
+    assert evaluate(x) == 3.5
+    assert evaluate(x.value()) == 3.5
+    assert evaluate(x.read_value()) == 3.5
+    assert evaluate(tf.identity(x)) == 3.5
+
+    with quantized_scope.scope(True):
+        assert evaluate(x) == 7
+        assert evaluate(x.value()) == 7
+        assert evaluate(x.read_value()) == 7
+        assert evaluate(tf.identity(x)) == 7
+
+
+def test_sparse_reads(eager_and_graph_mode):
+    x = create_quantized_variable(get_var([1.0, 2.0]), quantizer=lambda x: 2 * x)
+    evaluate(x.initializer)
+
+    assert evaluate(x.sparse_read([0])) == 1
+    assert evaluate(x.gather_nd([0])) == 1
+    with quantized_scope.scope(True):
+        assert evaluate(x.sparse_read([0])) == 2
+        assert evaluate(x.gather_nd([0])) == 2
+
+
+def test_read_nested_scopes(distribute_scope, eager_and_graph_mode):
+    x = create_quantized_variable(get_var(3.5), quantizer=lambda x: 2 * x)
+    evaluate(x.initializer)
+    with quantized_scope.scope(True):
+        assert evaluate(x.read_value()) == 7
+        with quantized_scope.scope(False):
+            assert evaluate(x.read_value()) == 3.5
+        assert evaluate(x.read_value()) == 7
+
+
+def test_method_delegations(distribute_scope, eager_and_graph_mode):
+    x = create_quantized_variable(get_var(3.5), quantizer=lambda x: 2 * x)
+    with quantized_scope.scope(True):
+        evaluate(x.initializer)
+        assert evaluate(x.value()) == 7
+        assert evaluate(x.read_value()) == 7
+        assert x.trainable
+        assert x.synchronization == x.latent_variable.synchronization
+        assert x.aggregation == x.latent_variable.aggregation
+        assert evaluate(x.initialized_value()) == 7
+        if not tf.executing_eagerly():
+            if not distribute_scope:
+                # These functions are not supported for DistributedVariables
+                x.load(4.5)
+                assert x.eval() == 9
+            assert evaluate(x.initial_value) == 7
+            assert x.op == x.latent_variable.op
+            assert x.graph == x.latent_variable.graph
[…]
