### 2019-12-08 33f5657
Improve argument validation
