### 2019-12-08 89f8dae
Fix snapshot tests
