+++ larq/snapshots/snap_quantized_variable_test.py
@@ -7,14 +7,14 @@ from snapshottest import Snapshot
 
 snapshots = Snapshot()
 
-snapshots['test_repr[eager] 1'] = "<QuantizedVariable 'Variable:0' shape=() dtype=float32 quantizer=<lambda> numpy=0.0>"
+snapshots['test_repr[eager] 1'] = "<QuantizedVariable 'x:0' shape=() dtype=float32 quantizer=<lambda> numpy=0.0>"
 
-snapshots['test_repr[eager] 2'] = "<QuantizedVariable 'Variable:0' shape=() dtype=float32 quantizer=Quantizer numpy=0.0>"
+snapshots['test_repr[eager] 2'] = "<QuantizedVariable 'x:0' shape=() dtype=float32 quantizer=Quantizer numpy=0.0>"
 
-snapshots['test_repr[eager] 3'] = "<QuantizedVariable 'Variable:0' shape=() dtype=float32 precision=1 numpy=0.0>"
+snapshots['test_repr[eager] 3'] = "<QuantizedVariable 'x:0' shape=() dtype=float32 precision=1 numpy=0.0>"
 
-snapshots['test_repr[graph] 1'] = "<QuantizedVariable 'Variable:0' shape=() dtype=float32 quantizer=<lambda>>"
+snapshots['test_repr[graph] 1'] = "<QuantizedVariable 'x:0' shape=() dtype=float32 quantizer=<lambda>>"
 
-snapshots['test_repr[graph] 2'] = "<QuantizedVariable 'Variable:0' shape=() dtype=float32 quantizer=Quantizer>"
+snapshots['test_repr[graph] 2'] = "<QuantizedVariable 'x:0' shape=() dtype=float32 quantizer=Quantizer>"
 
-snapshots['test_repr[graph] 3'] = "<QuantizedVariable 'Variable:0' shape=() dtype=float32 precision=1>"
+snapshots['test_repr[graph] 3'] = "<QuantizedVariable 'x:0' shape=() dtype=float32 precision=1>"
