### 2019-12-08 #1746 @sbrugman
Hi @lgeiger,

The paper has really benefitted from the revisions, good job. I have some suggestions:

- Bringing the power ... everyday lives => Bringing this technology outside of data centers is not a key point here, the same statement could hold for running the models in a data center. You are trying to say that deep-learning algorithms have this potential.

- reseach => research

- The references are somewhat sloppy. NIPS for example is cited multiple times inconsistently. I would advise to pick one format and stick with that. Not sure if you do, but don't blindly trust Google Scholar for bibtex. Here is the prefered bibtex for the ICLR paper: https://dblp.org/rec/bibtex/conf/iclr/AlizadehFLG19. You might be interested in reading: https://academic-notes.netlify.com/post/references/
