### 2019-12-08 ae9a3f9
Fix unit tests for TensorFlow 1.15
