### 2019-12-10 ba3c1d3
Fix unittests for TF 1.14
