### 2019-12-10 c509970
Add packaging as explicit test dependency
