+++ larq/quantized_variable_test.py
@@ -240,11 +240,11 @@ def test_checkpoint(tmp_path, eager_and_graph_mode):
 
 
 def test_invalid_wrapped_usage(distribute_scope):
-    with pytest.raises(ValueError, match="variable must be of type"):
+    with pytest.raises(ValueError, match="`variable` must be of type"):
         create_quantized_variable(tf.constant([1.0]))
-    with pytest.raises(ValueError, match="quantizer must be callable or None"):
+    with pytest.raises(ValueError, match="`quantizer` must be `callable` or `None`"):
         create_quantized_variable(get_var([1.0]), 1)
-    with pytest.raises(ValueError, match="precision must be of type integer or None"):
+    with pytest.raises(ValueError, match="`precision` must be of type `int` or `None`"):
         create_quantized_variable(get_var([1.0]), precision=1.0)
 
 
