### 2019-12-10 0af9cad
Move _get_name to staticmethod
