### 2019-12-10 897fbb6
Move quantized fixture to conftest.py
