### 2019-12-10 2576b1f
Fix paper typo (#351)
