### 2019-12-10 5c71505
Cite conference versions of the papers (#352)
