+++ paper/paper.bib
@@ -18,12 +18,17 @@
     booktitle    = {NIPS Autodiff Workshop}
 }
 @article{bmxnetv2,
-    title        = {Training Competitive Binary Neural Networks from Scratch},
     author       = {Joseph Bethge and Marvin Bornstein and Adrian Loy and Haojin Yang and Christoph Meinel},
-    year         = 2018,
-    journal      = {ArXiv e-prints},
-    archiveprefix = {arXiv},
-    eprint       = {1812.01965}
+    title        = {Training Competitive Binary Neural Networks from Scratch},
+    journal      = {CoRR},
+    volume       = {abs/1812.01965},
+    year         = {2018},
+    url          = {http://arxiv.org/abs/1812.01965},
+    archivePrefix = {arXiv},
+    eprint       = {1812.01965},
+    timestamp    = {Tue, 01 Jan 2019 15:01:25 +0100},
+    biburl       = {https://dblp.org/rec/bib/journals/corr/abs-1812-01965},
+    bibsource    = {dblp computer science bibliography, https://dblp.org}
 }
 @misc{dabnn,
     title        = {daBNN: A Super Fast Inference Framework for Binary Neural Networks on ARM devices},
@@ -31,21 +36,15 @@
     year         = 2019,
     eprint       = {arXiv:1908.05858}
 }
-@article{bop,
+@incollection{bop,
     title        = {Latent Weights Do Not Exist: Rethinking Binarized Neural Network Optimization},
-    author       = {{Helwegen}, Koen and {Widdicombe}, James and {Geiger}, Lukas and {Liu}, Zechun and {Cheng}, Kwang-Ting and {Nusselder}, Roeland},
+    author       = {Helwegen, Koen and Widdicombe, James and Geiger, Lukas and Liu, Zechun and Cheng, Kwang-Ting and Nusselder, Roeland},
     year         = 2019,
-    month        = jun,
-    journal      = {arXiv e-prints},
-    pages        = {arXiv:1906.02107},
-    url          = {https://arxiv.org/abs/1906.02107},
-    keywords     = {Computer Science - Machine Learning, Statistics - Machine Learning},
-    eid          = {arXiv:1906.02107},
-    archiveprefix = {arXiv},
-    eprint       = {1906.02107},
-    primaryclass = {cs.LG},
-    adsurl       = {https://ui.adsabs.harvard.edu/abs/2019arXiv190602107H},
-    adsnote      = {Provided by the SAO/NASA Astrophysics Data System}
+    booktitle    = {Advances in Neural Information Processing Systems 32},
+    publisher    = {Curran Associates, Inc.},
+    pages        = {7531--7542},
+    url          = {http://papers.nips.cc/paper/8971-latent-weights-do-not-exist-rethinking-binarized-neural-network-optimization.pdf},
+    editor       = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d\textquotesingle Alch\'{e}-Buc and E. Fox and R. Garnett}
 }
 @incollection{binarynet,
     title        = {Binarized Neural Networks},
@@ -74,7 +73,9 @@
     title        = {Bi-Real Net: Enhancing the Performance of 1-bit CNNs With Improved Representational Capability and Advanced Training Algorithm},
     author       = {Zechun Liu and Baoyuan Wu and Wenhan Luo and Xin Yang and Wei Liu and Kwang{-}Ting Cheng},
     year         = 2018,
+    month        = sep,
     journal      = {ECCV},
+    booktitle    = {The European Conference on Computer Vision (ECCV)},
     volume       = {abs/1808.00278},
     url          = {http://arxiv.org/abs/1808.00278},
     archiveprefix = {arXiv},
@@ -83,18 +84,16 @@
[…]
