### 2019-12-11 ef0101c
Bump altair from 3.3.0 to 4.0.0
Bumps [altair](https://github.com/altair-viz/altair) from 3.3.0 to 4.0.0.
- [Release notes](https://github.com/altair-viz/altair/releases)
- [Commits](https://github.com/altair-viz/altair/compare/v3.3.0...v4.0.0)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
