### 2019-12-12 3cbf456
:arrow_up: vega-lite@4
