+++ mkdocs.yml
@@ -81,7 +81,7 @@ markdown_extensions:
 extra_javascript:
   - "https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-chtml.js"
   - "https://cdn.jsdelivr.net/npm/vega@5"
-  - "https://cdn.jsdelivr.net/npm/vega-lite@3"
+  - "https://cdn.jsdelivr.net/npm/vega-lite@4"
   - "https://cdn.jsdelivr.net/npm/vega-embed@6"
 
 extra_css:
