### 2019-12-12 #1746 @lgeiger
Thanks for the comments, I tired to update the references to refer to the conference versions.

> the same statement could hold for running the models in a data center

I don't think that is entirely true. The examples of self-driving cars or autonomous drones are not really possible with data center based networks due to latency or privacy constraints.
