### 2019-12-12 #1746 @casperdcl
- (minor) not sure about authorship; it seems that if the larq team is fine with the current paper's author list then this should not be a problem - but I can't see any confirmation of this.
- the "statement of need" in the paper is titled "Encouraging reproducible research in `larq/zoo`. This sounds more like text than a title. It should be moved to the text and a more explicit title ("Statement of need") used instead. This will also make it more obvious that this section should appear earlier in the paper. Id' suggest right after the introduction - that way the introduction outlines why BNNs are useful, and then the statement of need outlines why this software in particular is useful.
- the statement of need also doesn't express the same sentiments as the documentation/website (https://larq.dev and https://larq.dev/about)
- (major) performance claims such as "64 times speed-up" should 1. have a literature citation and 2. come with proof that `larq` at least somewhat delivers on this theoretical promise.
