### 2019-12-15 d36e814
Cite NeurIPS paper instead of arXiv version
