### 2019-12-16 01743c0
Make HyperparameterScheduler compatible with CaseOptimizer
