### 2019-12-16 c3c63e1
RFC: Quantized Variables (#312)
* Support depthwise and separable layers

This is hopefully a first step to reduce duplication in the base layers

* Simplify model summary

Simplify model summary

* Remove quantized_latent_weights and quantizer arguments

since they are not needed anymore with the new API

* Move non_trainable_weights to base class

* Add initial version of quantized variable

* Add distribution scope pytest fixture

* Fix eval

We can't use the quantize decorator here, since eval would return a Tensor instead of an numpy array.

* Update larq/layers_base.py

Co-Authored-By: L
[…]
