+++ larq/quantized_variable.py
@@ -1,0 +1,311 @@
+"""Contains QuantizedVariable, a variable that can be quantized in the forward pass."""
+import tensorflow as tf
+from tensorflow.python.distribute.values import DistributedVariable
+from tensorflow.python.framework import ops
+from tensorflow.python.ops import resource_variable_ops
+
+from larq import quantized_scope
+
+
+class QuantizedVariable(tf.Variable):
+    """A Variable that can be quantized in the forward pass in applicable contexts."""
+
+    def __init__(self, variable, quantizer=None, precision=None):
+        """Creates an QuantizedVariable instance.
+
+        # Arguments
+        variable: A floating-point resource variable to wrap.
+        quantizer: An optional quantizer to transform the floating-point
+            variable to a fake quantized variable.
+        precision: An optional integer defining the precision of the quantized
+            variable. If `None`, `quantizer.precision` is used.
+        """
+        if not resource_variable_ops.is_resource_variable(variable):
+            raise ValueError(
+                "`variable` must be of type `tf.ResourceVariable`, "
+                f"but got `{type(variable)}`."
+            )
+        if not (quantizer is None or callable(quantizer)):
+            raise ValueError(
+                "`quantizer` must be `callable` or `None`, "
+                f"but got `{type(quantizer)}`."
+            )
+        if not (precision is None or type(precision) == int):
+            raise ValueError(
+                "`precision` must be of type `int` or `None`, "
+                f"but got `{type(precision)}`."
+            )
+        self.latent_variable = variable
+        self.quantizer = quantizer
+        self.precision = precision or getattr(quantizer, "precision", None)
+
+    def _quantize(self, value):
+        if self.quantizer and quantized_scope.should_quantize():
+            return self.quantizer(value)
+        return value
+
+    def value(self):
+        return self._quantize(self.latent_variable.value())
+
+    def read_value(self):
+        return self._quantize(self.latent_variable.read_value())
+
+    def numpy(self):
+        return self._quantize(self.latent_variable).numpy()
+
+    def sparse_read(self, *args, **kwargs):
+        return self._quantize(self.latent_variable.sparse_read(*args, **kwargs))
+
+    def gather_nd(self, *args, **kwargs):
+        return self._quantize(self.latent_variable.gather_nd(*args, **kwargs))
+
+    def __getattr__(self, name):
+        return getattr(self.latent_variable, name)
+
+    def _dense_var_to_tensor(self, *args, **kwargs):
+        return self._quantize(
+            self.latent_variable._dense_var_to_tensor(*args, **kwargs)
+        )
+
+    def eval(self, session=None):
+        return self._quantize(self.latent_variable).eval(session=session)
+
+    def initialized_value(self):
+        return self._quantize(self.latent_variable.initialized_value())
+
+    @property
+    def initial_value(self):
+        return self._quantize(self.latent_variable.initial_value)
+
+    def _should_act_as_resource_variable(self):
+        """Pass resource_variable_ops.is_resource_variable check."""
+        pass
+
+    @staticmethod
+    def _get_name(obj):
+        try:
+            return obj.__name__
+        except AttributeError:
+            return obj.__class__.__name__
+
+    def __repr__(self):
+        repr_ = (
+            f"<{self.__class__.__name__} '{self.name}' "
+            f"shape={self.shape} dtype={self.dtype.name}"
+        )
+        if self.quantizer is not None:
+            repr_ += f" quantizer={self._get_name(self.quantizer)}"
+        if self.precision is not None:
+            repr_ += f" precision={self.precision}"
[…]

[…]
