### 2019-12-16 b174591
Simplify assign operations in Bop
The use of tf.compat.v1.assign was only to make sure we execute the same code path as TF. I am not sure if this is necessary for TF 1.x. Let's see what CI thinks.
