### 2019-12-16 42cae5d
Add test for HyperparameterScheduler
