### 2019-12-16 6b74c18
Make optimizer argument optional
