+++ larq/callbacks.py
@@ -27,9 +27,12 @@ class HyperparameterScheduler(tf.keras.callbacks.Callback):
     verbose: int. 0: quiet, 1: update messages.
     """
 
-    def __init__(self, optimizer, schedule, hyperparameter, verbose=0):
+    def __init__(self, schedule, hyperparameter,optimizer=None, verbose=0):
         super(HyperparameterScheduler, self).__init__()
-        self.optimizer = optimizer
+        if optimizer is None:
+            self.optimizer = self.model.optimizer
+        else:
+            self.optimizer = optimizer
         self.schedule = schedule
         self.hyperparameter = hyperparameter
         self.verbose = verbose
