### 2019-12-17 93eb404
Use metaclass to allow direct instantiation of QuantizedVariable
This allows us to remove the `create_quantized_variable` function at the expense of a slightly more complicated implementation. Users can now directly instantiate `QantizedVariable` and don't have to worry about correctly handling distribution strategies.
