### 2019-12-17 21ef500
Use from_variable factory to create quantized variables
