+++ larq/quantized_variable_test.py
@@ -5,7 +5,7 @@ from packaging import version
 from tensorflow.python.distribute.values import DistributedVariable
 
 from larq import quantized_scope
-from larq.quantized_variable import QuantizedVariable, create_quantized_variable
+from larq.quantized_variable import QuantizedVariable
 from larq.testing_utils import evaluate
 
 
@@ -15,14 +15,14 @@ def get_var(val, dtype=None, name=None):
 
 def test_inheritance(distribute_scope):
     variable = get_var(3.0)
-    quantized_variable = create_quantized_variable(variable)
+    quantized_variable = QuantizedVariable.from_variable(variable)
     assert isinstance(quantized_variable, QuantizedVariable)
     assert isinstance(quantized_variable, tf.Variable)
     assert isinstance(quantized_variable, DistributedVariable) is distribute_scope  # type: ignore
 
 
 def test_read(distribute_scope, eager_and_graph_mode):
-    x = create_quantized_variable(get_var(3.5), quantizer=lambda x: 2 * x)
+    x = QuantizedVariable.from_variable(get_var(3.5), quantizer=lambda x: 2 * x)
     evaluate(x.initializer)
 
     assert evaluate(x) == 3.5
@@ -38,7 +38,7 @@ def test_read(distribute_scope, eager_and_graph_mode):
 
 
 def test_sparse_reads(eager_and_graph_mode):
-    x = create_quantized_variable(get_var([1.0, 2.0]), quantizer=lambda x: 2 * x)
+    x = QuantizedVariable.from_variable(get_var([1.0, 2.0]), quantizer=lambda x: 2 * x)
     evaluate(x.initializer)
 
     assert evaluate(x.sparse_read([0])) == 1
@@ -49,7 +49,7 @@ def test_sparse_reads(eager_and_graph_mode):
 
 
 def test_read_nested_scopes(distribute_scope, eager_and_graph_mode):
-    x = create_quantized_variable(get_var(3.5), quantizer=lambda x: 2 * x)
+    x = QuantizedVariable.from_variable(get_var(3.5), quantizer=lambda x: 2 * x)
     evaluate(x.initializer)
     with quantized_scope.scope(True):
         assert evaluate(x.read_value()) == 7
@@ -59,7 +59,7 @@ def test_read_nested_scopes(distribute_scope, eager_and_graph_mode):
 
 
 def test_method_delegations(distribute_scope, eager_and_graph_mode):
-    x = create_quantized_variable(get_var(3.5), quantizer=lambda x: 2 * x)
+    x = QuantizedVariable.from_variable(get_var(3.5), quantizer=lambda x: 2 * x)
     with quantized_scope.scope(True):
         evaluate(x.initializer)
         assert evaluate(x.value()) == 7
@@ -91,7 +91,7 @@ def test_method_delegations(distribute_scope, eager_and_graph_mode):
 
 
 def test_scatter_method_delegations(eager_and_graph_mode):
-    x = create_quantized_variable(get_var([3.5, 4]), quantizer=lambda x: 2 * x)
+    x = QuantizedVariable.from_variable(get_var([3.5, 4]), quantizer=lambda x: 2 * x)
     evaluate(x.initializer)
     with quantized_scope.scope(True):
         assert_array_equal(evaluate(x.value()), [7, 8])
@@ -120,9 +120,9 @@ def test_scatter_method_delegations(eager_and_graph_mode):
 
 def test_overloads(quantized, distribute_scope, eager_and_graph_mode):
     if quantized:
-        x = create_quantized_variable(get_var(3.5), quantizer=lambda x: 2 * x)
[…]
