### 2019-12-17 ed2970e
Make _maybe_wrap a staticmethod
