### 2019-12-18 3d42903
Change isort config to simplify imports
