+++ setup.cfg
@@ -12,6 +12,7 @@ line_length = 88
 use_parentheses = True
 multi_line_output = 3
 include_trailing_comma = True
+combine_as_imports = True
 skip =
     build
     dist
