### 2019-12-18 5b288d6
Use from_variable factory method to instantiate QuantizedVariables (#359)
* Use from_variable factory to create quantized variables

* Make _maybe_wrap a staticmethod

* Change isort config to simplify imports
