### 2019-12-18 6d01fc4
Apply suggestions from code review
