### 2019-12-18 583b346
Revert "Apply suggestions from code review"
This reverts commit 6d01fc411745a40c82e84f55429cc290081e3c93.
