### 2019-12-18 178c7ed
Make HyperparameterScheduler compatible with CaseOptimizer (#356)
* Make HyperparameterScheduler compatible with CaseOptimizer

* Black format code example

* Add test for HyperparameterScheduler

* Make optimizer argument optional

* black formatting

* Change order of imports

* Feedback

* Feedback2

* Apply suggestions from code review

* Revert "Apply suggestions from code review"

This reverts commit 6d01fc411745a40c82e84f55429cc290081e3c93.
