+++ larq/callbacks.py
@@ -4,26 +4,40 @@ import tensorflow as tf
 class HyperparameterScheduler(tf.keras.callbacks.Callback):
     """Generic hyperparameter scheduler.
 
+    !!! example
+        ```python
+        bop = lq.optimizers.Bop(threshold=1e-6, gamma=1e-3)
+        adam = tf.keras.optimizers.Adam(0.01)
+        optimizer = lq.optimizers.CaseOptimizer(
+            (lq.optimizers.Bop.is_binary_variable, bop), default_optimizer=adam,
+        )
+        callbacks = [
+            HyperparameterScheduler(lambda x: 0.001 * (0.1 ** (x // 30)), "gamma", bop)
+        ]
+        ```
     # Arguments
+    optimizer: the optimizer that contains the hyperparameter that will be scheduled.
+        Defaults to `self.model.optimizer` if `optimizer == None`.
     schedule: a function that takes an epoch index as input
         (integer, indexed from 0) and returns a new hyperparameter as output.
     hyperparameter: str. the name of the hyperparameter to be scheduled.
     verbose: int. 0: quiet, 1: update messages.
     """
 
-    def __init__(self, schedule, hyperparameter, verbose=0):
+    def __init__(self, schedule, hyperparameter, optimizer=None, verbose=0):
         super(HyperparameterScheduler, self).__init__()
+        self.optimizer = optimizer if optimizer else self.model.optimizer
         self.schedule = schedule
         self.hyperparameter = hyperparameter
         self.verbose = verbose
 
     def on_epoch_begin(self, epoch, logs=None):
-        if not hasattr(self.model.optimizer, self.hyperparameter):
+        if not hasattr(self.optimizer, self.hyperparameter):
             raise ValueError(
                 f'Optimizer must have a "{self.hyperparameter}" attribute.'
             )
 
-        hp = getattr(self.model.optimizer, self.hyperparameter)
+        hp = getattr(self.optimizer, self.hyperparameter)
         try:  # new API
             hyperparameter_val = tf.keras.backend.get_value(hp)
             hyperparameter_val = self.schedule(epoch, hyperparameter_val)
@@ -34,10 +48,10 @@ class HyperparameterScheduler(tf.keras.callbacks.Callback):
 
         if self.verbose > 0:
             print(
-                f"Epoch {epoch + 1}: {self.hyperparameter} changning to {tf.keras.backend.get_value(hp)}."
+                f"Epoch {epoch + 1}: {self.hyperparameter} changing to {tf.keras.backend.get_value(hp)}."
             )
 
     def on_epoch_end(self, epoch, logs=None):
         logs = logs or {}
-        hp = getattr(self.model.optimizer, self.hyperparameter)
+        hp = getattr(self.optimizer, self.hyperparameter)
         logs[self.hyperparameter] = tf.keras.backend.get_value(hp)
