### 2019-12-18 950ffd9
Change default hyperparameters of Bop
