+++ larq/layers.py
@@ -67,7 +67,9 @@ class QuantDense(QuantizerBase, tf.keras.layers.Dense):
         no activation is applied (`a(x) = x`).
     use_bias: Boolean, whether the layer uses a bias vector.
     input_quantizer: Quantization function applied to the input of the layer.
-    kernel_quantizer: Quantization function applied to the `kernel` weights matrix.
+    kernel_quantizer: Quantization function applied to the `kernel` weights matrix
+        or integer defining the desired precision if quantization should be handled
+        by an external optimizer.
     kernel_initializer: Initializer for the `kernel` weights matrix.
     bias_initializer: Initializer for the bias vector.
     kernel_regularizer: Regularizer function applied to the `kernel` weights matrix.
@@ -162,7 +164,9 @@ class QuantConv1D(QuantizerBase, tf.keras.layers.Conv1D):
         is applied (`a(x) = x`).
     use_bias: Boolean, whether the layer uses a bias vector.
     input_quantizer: Quantization function applied to the input of the layer.
-    kernel_quantizer: Quantization function applied to the `kernel` weights matrix.
+    kernel_quantizer: Quantization function applied to the `kernel` weights matrix
+        or integer defining the desired precision if quantization should be handled
+        by an external optimizer.
     kernel_initializer: Initializer for the `kernel` weights matrix.
     bias_initializer: Initializer for the bias vector.
     kernel_regularizer: Regularizer function applied to the `kernel` weights matrix.
@@ -270,7 +274,9 @@ class QuantConv2D(QuantizerBase, tf.keras.layers.Conv2D):
         no activation is applied (`a(x) = x`).
     use_bias: Boolean, whether the layer uses a bias vector.
     input_quantizer: Quantization function applied to the input of the layer.
-    kernel_quantizer: Quantization function applied to the `kernel` weights matrix.
+    kernel_quantizer: Quantization function applied to the `kernel` weights matrix
+        or integer defining the desired precision if quantization should be handled
+        by an external optimizer.
     kernel_initializer: Initializer for the `kernel` weights matrix.
     bias_initializer: Initializer for the bias vector.
     kernel_regularizer: Regularizer function applied to the `kernel` weights matrix.
@@ -385,7 +391,9 @@ class QuantConv3D(QuantizerBase, tf.keras.layers.Conv3D):
         no activation is applied (`a(x) = x`).
     use_bias: Boolean, whether the layer uses a bias vector.
     input_quantizer: Quantization function applied to the input of the layer.
-    kernel_quantizer: Quantization function applied to the `kernel` weights matrix.
+    kernel_quantizer: Quantization function applied to the `kernel` weights matrix
+        or integer defining the desired precision if quantization should be handled
+        by an external optimizer.
     kernel_initializer: Initializer for the `kernel` weights matrix.
     bias_initializer: Initializer for the bias vector.
     kernel_regularizer: Regularizer function applied to the `kernel` weights matrix.
@@ -494,7 +502,8 @@ class QuantDepthwiseConv2D(QuantizerDepthwiseBase, tf.keras.layers.DepthwiseConv
     use_bias: Boolean, whether the layer uses a bias vector.
     input_quantizer: Quantization function applied to the input of the layer.
     depthwise_quantizer: Quantization function applied to the `depthwise_kernel`
-        weights matrix.
+        weights matrix or integer defining the desired precision if quantization should
+        be handled by an external optimizer.
     depthwise_initializer: Initializer for the depthwise kernel matrix.
     bias_initializer: Initializer for the bias vector.
     depthwise_regularizer: Regularizer function applied to the depthwise kernel matrix.
@@ -597,8 +606,12 @@ class QuantSeparableConv1D(QuantizerSeparableBase, tf.keras.layers.SeparableConv
[…]
