### 2019-12-19 dfb1fe6
rename "external optimizer" -> "specialized optimizer"
