### 2019-12-19 9d07af9
Improve warning in Bop
