### 2019-12-19 9ba97e4
Improve docstring of Bop
