### 2019-12-22 9bd4b44
Bump pymdown-extensions from 6.2 to 6.2.1 (#366)
Bumps [pymdown-extensions](https://github.com/facelessuser/pymdown-extensions) from 6.2 to 6.2.1.
- [Release notes](https://github.com/facelessuser/pymdown-extensions/releases)
- [Commits](https://github.com/facelessuser/pymdown-extensions/compare/6.2.0...6.2.1)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
