### 2019-12-30 48a9437
Remove outdated line in Bop documentation (#367)
