### 2020-01-02 28f8fb9
Bump pillow from 6.2.1 to 7.0.0
Bumps [pillow](https://github.com/python-pillow/Pillow) from 6.2.1 to 7.0.0.
- [Release notes](https://github.com/python-pillow/Pillow/releases)
- [Changelog](https://github.com/python-pillow/Pillow/blob/master/CHANGES.rst)
- [Commits](https://github.com/python-pillow/Pillow/compare/6.2.1...7.0.0)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
