### 2020-01-02 96e33c7
Fix weight_clip documentation issue
Since, `lq.constraints.weight_clip` is a alias to `lq.constraints.WeightClip` similar to how [upstream constraints](https://github.com/tensorflow/tensorflow/blob/5f8de9502dcf0bfb769a1df296b95b1241fc4d16/tensorflow/python/keras/constraints.py#L254-L258) are implemented. This means `lq.constraints.weight_clip` is actually a `class` and therfor needs to be instantiated first (e.g. `lq.constraints.weight_clip()`). Unfortunately TensorFlow throws a really confusing error message, so this is not obvious.
