### 2020-01-03 96d3312
Address paper review comments (#370)
* Paper: Improve introduction and clarify performance claims

* Add Statement of need section
