### 2020-01-03 #1746 @lgeiger
Happy 2020 everyone 🎉

> Let me clarify. The introduction appears unclear to me, when reading it for the first time.

@sbrugman Thanks for clarifying. I rewrote the introduction to make it less confusing.

> the "statement of need" in the paper is titled "Encouraging reproducible research in `larq/zoo`"

@casperdcl That is a good point, I haven't though about that. I consolidated the spread out statement of need into its own section.

> (major) performance claims such as "64 times speed-up" should 1. have a literature citation and 2. come with proof that larq at least somewhat delivers on this theoretical promise.

@casperdcl Thanks for catching this, indeed performance claims like this can be very misleading. The "64 times speed-up" stems from the `n^2` logic gates per `n` bits in classic multiplier circuits. I agree, this claim is problematic and would need real world measurements to confirm. I changed it to `Compared to an equivalent 8-bit quantized network BNNs require 8 times smaller
memory size and 8 times fewer memory accesses,...` which is cited from Hubara et al., 2016.
