### 2020-01-08 3f51916
Update pytype requirement
Updates the requirements on [pytype](https://github.com/google/pytype) to permit the latest version.
- [Release notes](https://github.com/google/pytype/releases)
- [Changelog](https://github.com/google/pytype/blob/master/CHANGELOG)
- [Commits](https://github.com/google/pytype/compare/2019.10.17...2020.01.07)

Signed-off-by: dependabot-preview[bot] <support@dependabot.com>
