+++ setup.py
@@ -34,7 +34,7 @@ setup(
             "pytest>=5.2.4,<5.4.0",
             "pytest-cov~=2.8.1",
             "pytest-xdist>=1.30,<1.32",
-            "pytype>=2019.10.17,<2019.13.0",
+            "pytype>=2019.10.17,<2020.2.0",
             "snapshottest~=0.5.1",
         ],
         "docs": [
