+++ requirements.txt
@@ -1,4 +1,4 @@
-tensorflow==2.0.0
+tensorflow==2.1.0
 tensorflow_datasets==1.3.2
 larq_zoo==0.5.0
 Pillow==7.0.0
