### 2020-01-15 #1746 @terrytangyuan
It looks like all items in the checklists are completed. Thanks everyone!

@lgeiger Could you fix the above missing DOIs?
