### 2020-01-15 a3ff5fb
Add missing DOIs to paper references
Fixes https://github.com/openjournals/joss-reviews/issues/1746#issuecomment-574671056
