### 2020-01-15 630618e
Add missing DOIs to paper references (#383)
Fixes https://github.com/openjournals/joss-reviews/issues/1746#issuecomment-574671056
