+++ paper/paper.bib
@@ -69,20 +69,17 @@
     biburl       = {https://dblp.org/rec/bib/journals/corr/abs-1906-08637},
     bibsource    = {dblp computer science bibliography, https://dblp.org}
 }
-@article{bireal_net,
+@inproceedings{bireal_net,
     title        = {Bi-Real Net: Enhancing the Performance of 1-bit CNNs With Improved Representational Capability and Advanced Training Algorithm},
     author       = {Zechun Liu and Baoyuan Wu and Wenhan Luo and Xin Yang and Wei Liu and Kwang{-}Ting Cheng},
     year         = 2018,
-    month        = sep,
-    journal      = {ECCV},
-    booktitle    = {The European Conference on Computer Vision (ECCV)},
-    volume       = {abs/1808.00278},
-    url          = {http://arxiv.org/abs/1808.00278},
-    archiveprefix = {arXiv},
-    eprint       = {1808.00278},
-    timestamp    = {Thu, 12 Sep 2019 15:28:48 +0200},
-    biburl       = {https://dblp.org/rec/bib/journals/corr/abs-1808-00278},
-    bibsource    = {dblp computer science bibliography, https://dblp.org}
+    booktitle    = {Computer Vision -- ECCV 2018},
+    publisher    = {Springer International Publishing},
+    address      = {Cham},
+    pages        = {747--763},
+    isbn         = {978-3-030-01267-0},
+    doi          = {10.1007/978-3-030-01267-0_44},
+    editor       = {Ferrari, Vittorio and Hebert, Martial and Sminchisescu, Cristian and Weiss, Yair}
 }
 @inproceedings{xnor_net,
     title        = {XNOR-Net: ImageNet Classification Using Binary Convolutional Neural Networks},
@@ -93,6 +90,7 @@
     address      = {Cham},
     pages        = {525--542},
     isbn         = {978-3-319-46493-0},
+    doi          = {10.1007/978-3-319-46493-0_32},
     editor       = {Leibe, Bastian and Matas, Jiri and Sebe, Nicu and Welling, Max}
 }
 @article{dorefa,
@@ -148,25 +146,26 @@
     biburl       = {https://dblp.org/rec/bib/conf/iclr/AlizadehFLG19},
     bibsource    = {dblp computer science bibliography, https://dblp.org}
 }
-@article{Zhu2018,
+@inproceedings{Zhu2018,
     author       = {Shilin Zhu and Xin Dong and Hao Su},
     title        = {Binary Ensemble Neural Network: More Bits per Network or More Networks per Bit?},
-    journal      = {CoRR},
-    volume       = {abs/1806.07550},
-    year         = {2018},
-    url          = {http://arxiv.org/abs/1806.07550},
-    archivePrefix = {arXiv},
-    eprint       = {1806.07550},
-    timestamp    = {Mon, 13 Aug 2018 16:47:13 +0200},
-    biburl       = {https://dblp.org/rec/bib/journals/corr/abs-1806-07550},
-    bibsource    = {dblp computer science bibliography, https://dblp.org}
+    booktitle    = {2019 IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR)},
+    year         = 2019,
+    pages        = {4918-4927},
+    url          = {http://openaccess.thecvf.com/content_CVPR_2019/papers/Zhu_Binary_Ensemble_Neural_Network_More_Bits_per_Network_or_More_CVPR_2019_paper.pdf}
+    doi          = {10.1109/CVPR.2019.00506},
+    ISSN         = {1063-6919},
+    month        = {June}
 }
 @inproceedings{Zhuang2018,
[…]
