### 2020-01-15 #1746 @lgeiger
Thanks for the great reviews everyone!

> Could you fix the above missing DOIs?

@terrytangyuan I fixed the missing DOIs. Note that "Keras" doesn't have a DOI and the one linked by whedon is not correct.
