### 2020-01-15 #1746 @lgeiger
Great! I published [larq@v0.8.3](https://github.com/larq/larq/releases/tag/v0.8.3) with the changes from this review and uploaded it to figshare: https://doi.org/10.6084/m9.figshare.11619912.v1
