### 2020-01-15 #1746 @kyleniemeyer
Hi @lgeiger, I'm doing some final checks on the article before publishing.

It looks like the Zhang 2019 reference is missing some details—I just see the authors, year, and title. Where/how was this published?
