### 2020-01-15 8abe8ed
Change DABNN citation to cite conference version
Fixes https://github.com/openjournals/joss-reviews/issues/1746#issuecomment-574895592
