### 2020-01-15 232596d
Change DABNN citation to cite conference version (#386)
Fixes https://github.com/openjournals/joss-reviews/issues/1746#issuecomment-574895592
