+++ paper/paper.bib
@@ -30,11 +30,14 @@
     biburl       = {https://dblp.org/rec/bib/journals/corr/abs-1812-01965},
     bibsource    = {dblp computer science bibliography, https://dblp.org}
 }
-@misc{dabnn,
+@article{dabnn,
     title        = {daBNN: A Super Fast Inference Framework for Binary Neural Networks on ARM devices},
-    author       = {Jianhao Zhang and Yingwei Pan and Ting Yao and He Zhao and Tao Mei},
-    year         = 2019,
-    eprint       = {arXiv:1908.05858}
+    isbn         = {9781450368896},
+    doi          = {10.1145/3343031.3350534},
+    journal      = {Proceedings of the 27th ACM International Conference on Multimedia},
+    publisher    = {ACM Press},
+    author       = {Zhang, Jianhao and Pan, Yingwei and Yao, Ting and Zhao, He and Mei, Tao},
+    year         = 2019
 }
 @incollection{bop,
     title        = {Latent Weights Do Not Exist: Rethinking Binarized Neural Network Optimization},
