### 2020-01-15 #1746 @lgeiger
> It looks like the Zhang 2019 reference is missing some details—I just see the authors, year, and title. Where/how was this published?

@kyleniemeyer Good catch. I used the BibTeX entry mentioned in their GitHub repo which was missing the details. I switched to citing the article published at ACM now.
